% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{ark_node}
\alias{ark_node}
\title{Query or set the ARK node to use}
\usage{
ark_node(node = "https://api.arknode.net")
}
\arguments{
\item{node}{optional ARK node with HTTPS prefix}
}
\value{
string
}
\description{
Query or set the ARK node to use
}
