% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R, R/backtest_prep.R
\name{backtest}
\alias{backtest}
\title{conduct backtest}
\usage{
backtest(
  data = data,
  indicator = settings$indicator,
  initial_value = settings$initial_value,
  starting_date = settings$start_time,
  end_date = settings$end_time,
  min_trade = settings$min_trade,
  min_momentum_count = settings$min_momentum_count,
  fees = 0.001
)

backtest(
  data = data,
  indicator = settings$indicator,
  initial_value = settings$initial_value,
  starting_date = settings$start_time,
  end_date = settings$end_time,
  min_trade = settings$min_trade,
  min_momentum_count = settings$min_momentum_count,
  fees = 0.001
)
}
\arguments{
\item{fees}{}
}
\description{
conduct backtest

conduct backtest
}
