% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest_settings_check}
\alias{backtest_settings_check}
\title{Check the settings specified for backtesting}
\usage{
backtest_settings_check()
}
\description{
Check the settings specified for backtesting
}
