% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_account}
\alias{binance_account}
\title{Get current general Binance account information, without balances}
\usage{
binance_account()
}
\value{
data.table
}
\description{
Get current general Binance account information, without balances
}
