% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_account.R
\name{binance_account_balance}
\alias{binance_account_balance}
\title{Binance Account Balances}
\usage{
binance_account_balance()
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Get current balances of the connected account.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
binance_credentials("api-key", "api-secret")
binance_account_balance()
\dontshow{\}) # examplesIf}
}
\keyword{AccountEndpoints}
