% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_account.R
\name{binance_account_trades}
\alias{binance_account_trades}
\title{Binance Account Trades}
\usage{
binance_account_trades(pair, from, to, from_id, limit, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{from}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical trades.
If it is \code{missing}, the default, retrieved the last 1000 trades from \code{to} date.}

\item{to}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical trades.
If it is \code{missing}, the default, will be used as end date \code{\link[=Sys.time]{Sys.time()}}.}

\item{from_id}{Numeric. The last id from which retrieve the trades. Default is \code{missing}.}

\item{limit}{Integer. The maximum number of trades to retrieve.
If \code{missing}, the default, will be used the maximum value that is 1000.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Get trades for a specific account and symbol.
}
\keyword{AccountEndpoints}
