% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_all_orders}
\alias{binance_all_orders}
\title{Fetch all orders from the Binance account}
\usage{
binance_all_orders(symbol, order_id, start_time, end_time, limit)
}
\arguments{
\item{symbol}{string}

\item{order_id}{optional number}

\item{start_time}{optional POSIX timestamp}

\item{end_time}{optional POSIX timestamp}

\item{limit}{optional int}
}
\value{
data.table
}
\description{
Fetch all orders from the Binance account
}
\examples{
\dontrun{
binance_cancel_order('ARKETH', 8)
binance_cancel_order('ARKBTC', 'myOrder7')
}
}
