% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{binance_balances}
\alias{binance_balances}
\title{Get current Binance balances in a nice table}
\usage{
binance_balances(threhsold = 0)
}
\arguments{
\item{threhsold}{filter out symbols with less then this number of coins}
}
\value{
data.table
}
\description{
Get current Binance balances in a nice table
}
