% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_book_ticker.R
\name{binance_book_ticker}
\alias{binance_book_ticker}
\title{Symbol Order Book Ticker}
\usage{
binance_book_ticker(pair, api, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}. Multiple pairs are allowed only if \code{api} is \code{"spot"}.
If \code{missing}, the default, will be returned the book ticker for all the trading pairs.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for endpoint \href{https://developers.binance.com/docs/binance-spot-api-docs/rest-api/market-data-endpoints#symbol-order-book-ticker}{api/v3/ticker/bookTicker}.
The ip weight is 2 if a symbol is submitted, otherwise is 4. The maximum ip weight is 200.
The ip weight depends on the number of pairs requested. The maximum ip weight is 80.
\item \code{"fapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Symbol-Order-Book-Ticker}{fapi/v1/ticker/bookTicker}.
The ip weight is 2 if a symbol is submitted, otherwise is 5.
\item \code{"dapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api/Symbol-Order-Book-Ticker}{dapi/v1/ticker/bookTicker}.
The ip weight is 2 if a symbol is submitted, otherwise is 5.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[=data.frame-class]{data.frame}} with 8 columns:
\itemize{
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, time of the snapshot.
\item \code{pair}: Character, reference trading pair, present only if \code{api} is \code{"dapi"}.
\item \code{symbol}: Character, trading pair.
\item \code{market}: Character, reference API.
\item \code{ask}: Numeric, best ASK price.
\item \code{bid}: Numeric, best BID price.
\item \code{ask_quantity}: Numeric, quantity at best ASK price.
\item \code{bid_quantity}: Numeric, quantity at best BID price.
}
}
\description{
Get best ASK and BID price and quantities on the order book for a pair or a list of pairs.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Get book ticker for all pairs
binance_book_ticker(api = "spot")
binance_book_ticker(api = "fapi")
binance_book_ticker(api = "dapi")

# Get book ticker for BTCUSDT
binance_book_ticker(pair = "BTCUSDT", api = "spot")
\dontshow{\}) # examplesIf}
}
\keyword{marketEndpoints}
