% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_trading.R
\name{binance_cancel_order}
\alias{binance_cancel_order}
\title{Binance Cancel Order}
\usage{
binance_cancel_order(pair, order_id, client_order_id)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{order_id}{Numeric. Order id that uniquely identify the trade.}

\item{client_order_id}{Numeric. Client order id that uniquely identify the trade.}
}
\value{
A \code{\link[tibble]{tibble}} with canceled order details.
}
\description{
Cancel an active order.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
binance_credentials("api-key", "api-secret")
binance_cancel_order(pair = "BTCUSDT", order_id = 123456)
\dontshow{\}) # examplesIf}
}
\keyword{TradingEndpoints}
