% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{binance_coins}
\alias{binance_coins}
\title{Get all currently valid coin names from Binance}
\usage{
binance_coins()
}
\value{
character vector
}
\description{
Get all currently valid coin names from Binance
}
