% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{binance_coins_prices}
\alias{binance_coins_prices}
\title{Get all currently valid coin names from Binance along with the USDT prices}
\usage{
binance_coins_prices(unit = "USDT")
}
\value{
data.table
}
\description{
Get all currently valid coin names from Binance along with the USDT prices
}
