% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_account-internals.R
\name{binance_credentials}
\alias{binance_credentials}
\title{Set Binance API Keys}
\usage{
binance_credentials(key, secret)
}
\arguments{
\item{key}{Character scalar. API key. If missing or \code{NULL}, the stored key is removed.}

\item{secret}{Character scalar. API secret. If missing or \code{NULL}, the stored secret is removed.}
}
\value{
Invisibly returns \code{NULL} after updating credentials in the package namespace.
}
\description{
Sets the API key and secret to interact with authenticated Binance API endpoints.
}
\examples{
\dontrun{
# Add api keys 
binance_credentials('foo', 'bar')
# Remove api keys 
binance_credentials()
}

}
\keyword{AccountEndpoints}
