% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_depth.R
\name{binance_depth}
\alias{binance_depth}
\title{Binance Order Book Snapshot}
\usage{
binance_depth(pair, api, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for endpoint \href{https://developers.binance.com/docs/binance-spot-api-docs/rest-api/market-data-endpoints#order-book}{api/v3/depth}. The ip weight is 250.
\item \code{"fapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Order-Book}{fapi/v1/depth}. The ip weight is 20.
\item \code{"dapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api/Order-Book}{dapi/v1/depth}. The ip weight is 20.
\item \code{"eapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/option/market-data/Order-Book}{eapi/v1/depth}. The ip weight is 1.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}} with 7 columns:
\itemize{
\item \code{last_update_id}: Integer, id of the last snapshot.
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, time of the snapshot.
\item \code{market}: Character, reference API.
\item \code{pair}: Character, trading pair.
\item \code{side}: Character, side of the limit orders in the book. Can be \code{"ASK"} or \code{"BID"}.
\item \code{price}: Numeric, price level.
\item \code{quantity}: Numeric, quantity for each price level.
}
}
\description{
Retrieve a snapshot of the order book for a trading pair.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Get the order book for BTCUSDT
binance_depth(pair = "BTCUSDT", api = "spot")
binance_depth(pair = "BTCUSDT", api = "fapi")

# Get the order book for BTCUSD_PERP
binance_depth(pair = "BTCUSD_PERP", api = "dapi")

# Get the order book for a put option on BTC 
binance_depth(pair = "BTC-240628-30000-P", api = "eapi")
\dontshow{\}) # examplesIf}
}
\keyword{marketEndpoints}
