% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_dust_assets}
\alias{binance_dust_assets}
\title{Fetch all assets that can be converted to dust in the Binance account}
\usage{
binance_dust_assets(max_btc = 1e-04)
}
\value{
data.table
}
\description{
Fetch all assets that can be converted to dust in the Binance account
}
