% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_dust_conversion}
\alias{binance_dust_conversion}
\title{Convert 'dust' assets in the Binance account into BNB}
\usage{
binance_dust_conversion(dust)
}
\arguments{
\item{dust}{The data.table output of binance_dust_assets function}
}
\value{
data.table
}
\description{
Convert 'dust' assets in the Binance account into BNB
}
