% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_exchange_info.R
\name{binance_exchange_info}
\alias{binance_exchange_info}
\title{Binance Market Information}
\usage{
binance_exchange_info(api, permissions, pair, quiet = FALSE)
}
\arguments{
\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for endpoint \href{https://developers.binance.com/docs/binance-spot-api-docs/rest-api/general-endpoints#exchange-information}{api/v3/exchangeInfo}. The ip weight is 20.
\item \code{"fapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Exchange-Information}{fapi/v1/exchangeInfo}. The ip weight is 1.
\item \code{"dapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api/Exchange-Information}{dapi/v1/exchangeInfo}. The ip weight is 1.
\item \code{"eapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/option/market-data/Exchange-Information}{eapi/v1/exchangeInfo}. The ip weight is 1.
}}

\item{permissions}{Character, optional. Used only if \code{api = "spot"}. Types of trading pairs to return.
If \code{"all"}, the default, all available types of pairs will be returned. Available options are:
\itemize{
\item \code{"all"}: trading pairs in all markets.
\item \code{"spot"}: trading pairs only in spot markets.
\item \code{"margin"}: trading pairs only in margin markets.
\item \code{"leveraged"}: trading pairs only in leveraged markets.
}}

\item{pair}{Character, optional. Trading pair, e.g. \code{"BTCUSDT"}. If \code{NULL}, the default, all available pairs will be returned.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Obtain market information and available trading pairs.
}
\examples{
# Get all pairs in all markets
binance_exchange_info(api = "spot", permissions = "all", pair = NULL)

# Get all pairs only in spot markets
binance_exchange_info(api = "spot", permissions = "spot", pair = NULL)

# Get all pairs only in margin markets
binance_exchange_info(api = "spot", permissions = "margin", pair = NULL)

# Get all pairs only in leveraged market
binance_exchange_info(api = "spot", permissions = "leveraged", pair = NULL)

# Get information only for BTCUSDT in all markets
binance_exchange_info(api = "spot", permissions = "all", pair = "BTCUSDT")

# Get information for multiple pairs in all markets
binance_exchange_info(api = "spot", 
                      permissions = "all", 
                      pair = c("BTCUSDT", "BNBUSDT"))

# Get information for multiple pairs only in margin and leveraged markets
binance_exchange_info(api = "spot", 
                      permissions = c("margin", "leveraged"), 
                      pair = c("BTCBUSD", "ETHBUSD"))
                      
# Get all pairs in futures USD-m markets
binance_exchange_info(api = "fapi")

# Get all pairs in futures COIN-m markets
binance_exchange_info(api = "dapi")

# Get all pairs in options markets
binance_exchange_info(api = "eapi")

}
\keyword{generalEndpoints}
