% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_filters}
\alias{binance_filters}
\title{Get current filters for a symbol}
\usage{
binance_filters(symbol)
}
\arguments{
\item{symbol}{string}
}
\value{
data.table
}
\description{
Get current filters for a symbol
}
