% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_futures_stats.R
\name{binance_futures_stats}
\alias{binance_futures_stats}
\title{Futures Statistics}
\usage{
binance_futures_stats(pair, api, interval, indicator, from, to, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"} or \code{"BTCUSD"}.}

\item{api}{Character, reference API. Available options are \code{"fapi"} or \code{"dapi"}.}

\item{interval}{Character. Default is \code{"1h"}. Time interval for open interest data. Available intervals are:
\itemize{
\item Minutely: \code{"5m"}, \code{"15m"} and \code{"30m"}.
\item Hourly: \code{"1h"}, \code{"2h"}, \code{"4h"}, \code{"6h"}, \code{"8h"} and \code{"12h"}.
\item Daily: \code{"1d"}.
}}

\item{indicator}{Character reference statistic. Available indicators are:
\itemize{
\item \code{takerlongshortRatio}: only when \code{api = "fapi"}, for the endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Taker-BuySell-Volume}{futures/data/takerlongshortRatio}.
Taker Buy/Sell Volume. Taker Buy Volume: the total volume of buy orders filled by takers within the period. Taker Sell Volume: the total volume of sell orders filled by takers within the period.
\item \code{takerBuySellVol}: only when \code{api = "dapi"}, for the endpoint \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api/Taker-Buy-Sell-Volume}{futures/data/takerBuySellVol}.
Taker Buy/Sell Volume.
\item \code{globalLongShortAccountRatio}: for the endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Long-Short-Ratio}{futures/data/globalLongShortAccountRatio}.
The proportion of net long and net short positions to total open positions.
\item \code{topLongShortPositionRatio}: for the endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Top-Trader-Long-Short-Ratio}{futures/data/topLongShortPositionRatio}.
The proportion of net long and net short positions to total open positions of the top 20\% users with the highest margin balance.
\item \code{topLongShortAccountRatio}: for the endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Top-Long-Short-Account-Ratio}{futures/data/topLongShortAccountRatio}.
The proportion of net long and net short accounts to total accounts of the top 20\% users with the highest margin balance. Each account is counted once only.
}}

\item{from}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical data, only last 30 days are available.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()-lubridate::days(30)}.}

\item{to}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical data, only last 30 days are available.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()}.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
Get the historical statistics for futures.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
The historical open interest data are only available for the last 30 days.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Statistics in USD-m market     
binance_futures_stats(pair = "BTCUSDT", 
                      api = "fapi", 
                      interval = "1d", 
                      indicator = "takerlongshortRatio", 
                      from = Sys.Date()-2)
binance_futures_stats(pair = "BTCUSDT", 
                      api = "fapi", 
                      interval = "1d", 
                      indicator = "globalLongShortAccountRatio", 
                      from = Sys.Date()-2)
binance_futures_stats(pair = "BTCUSDT", 
                      api = "fapi", 
                      interval = "1d", 
                      indicator = "topLongShortPositionRatio", 
                      from = Sys.Date()-2)
binance_futures_stats(pair = "BTCUSDT", 
                      api = "fapi", 
                      interval = "1d", 
                      indicator = "topLongShortAccountRatio", 
                      from = Sys.Date()-2)    
                      
# Statistics in COIN-m market                       
binance_futures_stats(pair = "BTCUSD", 
                      api = "dapi", 
                      interval = "1d", 
                      indicator = "takerBuySellVol", 
                      from = Sys.Date()-2)
binance_futures_stats(pair = "BTCUSD", 
                      api = "dapi", 
                      interval = "1d", 
                      indicator = "globalLongShortAccountRatio", 
                      from = Sys.Date()-2)
binance_futures_stats(pair = "BTCUSD", 
                      api = "dapi", 
                      interval = "1d", 
                      indicator = "topLongShortPositionRatio", 
                      from = Sys.Date()-2)
binance_futures_stats(pair = "BTCUSD", 
                      api = "dapi", 
                      interval = "1d", 
                      indicator = "topLongShortAccountRatio", 
                      from = Sys.Date()-2) 
\dontshow{\}) # examplesIf}
}
\keyword{marketEndpoints}
