% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_get_destination}
\alias{binance_get_destination}
\title{check which destination is active}
\usage{
binance_get_destination()
}
\value{
the destination
}
\description{
check which destination is active
}
