% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_get_filters}
\alias{binance_get_filters}
\title{Compute current filters for a symbol}
\usage{
binance_get_filters(symbol)
}
\arguments{
\item{symbol}{string}
}
\value{
named vector
}
\description{
Compute current filters for a symbol
}
