% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_key}
\alias{binance_key}
\title{Look up Binance API key stored in the environment}
\usage{
binance_key()
}
\value{
string
}
\description{
Look up Binance API key stored in the environment
}
\keyword{internal}
