% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_last_open_interest.R
\name{binance_last_open_interest}
\alias{binance_last_open_interest}
\title{Binance Open Interest}
\usage{
binance_last_open_interest(pair, api, expiration, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"fapi"}. Available options are:
\itemize{
\item \code{"fapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Open-Interest}{fapi/v1/openInterest}. The ip weight is 1.
\item \code{"dapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api/Open-Interest}{dapi/v1/openInterest}. The ip weight is 1.
\item \code{"eapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/option/market-data/Open-Interest}{eapi/v1/openInterest}. The ip weight is 0.
}}

\item{expiration}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Used only if \code{api = "eapi"}.
Expiration date for options contracts. If it is \code{missing}, the default, will be used \code{\link[=Sys.time]{Sys.time()}}.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}} with 4 columns:
\itemize{
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, observation date.
\item \code{market}: Character, API.
\item \code{pair}: Character, trading pair.
\item \code{open_interest}: Numeric, open interest in base currency.
}
}
\description{
Get the current open interest data for a trading pair.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Get the open interest data for BTCUSDT
binance_last_open_interest(pair = "BTCUSDT", api = "fapi")

# Get the open interest data for BTCUSD_PERP
binance_last_open_interest(pair = "BTCUSD_PERP", api = "dapi")

# Get the open interest data for options on BTC
binance_last_open_interest(pair = "BTC", api = "eapi", expiration = Sys.Date() + 1)
\dontshow{\}) # examplesIf}
}
\keyword{marketEndpoints}
