% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_last_trades.R
\name{binance_last_trades}
\alias{binance_last_trades}
\title{Get Recent Trades}
\usage{
binance_last_trades(pair, api, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for endpoint \href{https://developers.binance.com/docs/binance-spot-api-docs/rest-api/market-data-endpoints#recent-trades-list}{api/v3/trades}. The ip weight is 25.
\item \code{"fapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Recent-Trades-List}{fapi/v1/trades}. The ip weight is 5.
\item \code{"dapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api/Recent-Trades-List}{dapi/v1/trades}. The ip weight is 5.
\item \code{"eapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/option/market-data/Recent-Trades-List}{eapi/v1/trades}. The ip weight is 5.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}} with 7 columns:
\itemize{
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, trade execution date.
\item \code{market}: Character, selected API.
\item \code{pair}: Character, trading pair.
\item \code{price}: Numeric, trade price.
\item \code{quantity}: Numeric, trade quantity.
\item \code{side}: Character, trade side. Can be \code{"BUY"} or \code{"SELL"}.
\item \code{trade_id}: Integer, trade id.
}
}
\description{
Get the last 1000 trades for a trading pair.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Get last 1000 trades for BTCUSDT
binance_last_trades(pair = "BTCUSDT", api = "spot")
binance_last_trades(pair = "BTCUSDT", api = "fapi")

# Get last 1000 trades for BTCUSD_PERP
binance_last_trades(pair = "BTCUSD_PERP", api = "dapi")

# Get last 1000 trades for a put option on BTC
binance_last_trades(pair = "BTC-240628-30000-P", api = "eapi")
\dontshow{\}) # examplesIf}
}
\keyword{marketEndpoints}
