% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws_live.R
\name{binance_live_depth}
\alias{binance_live_depth}
\title{Binance Live Depth}
\usage{
binance_live_depth(pair, api, quiet = FALSE)
}
\arguments{
\item{pair}{Character, trading pair, e.g. "BTCUSDT".}

\item{api}{Character, reference API. Available options are:
\itemize{
\item \code{spot}: For \href{https://binance-docs.github.io/apidocs/spot/en/#diff-depth-stream}{Spot API}.
\item \code{fapi}: For \href{https://binance-docs.github.io/apidocs/futures/en/#diff-book-depth-streams}{Futures USD-m API}.
}}

\item{quiet}{Logical, suppress informational messages if \code{TRUE}. Default \code{FALSE}.}
}
\description{
Live depth stream for a trading pair.
}
