% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws_live.R
\name{binance_live_trades}
\alias{binance_live_trades}
\title{Binance Live Trades}
\usage{
binance_live_trades(pair, api, from, to, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://binance-docs.github.io/apidocs/spot/en/#aggregate-trade-streams}{spot API}.
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#aggregate-trade-streams}{futures USD-M API}.
}}

\item{from}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical data.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()-lubridate::minutes(10)}.}

\item{to}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical data.
If it is \code{missing}, the default, will be used as end date \code{\link[=Sys.time]{Sys.time()}}.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\description{
Live trades stream for a trading pair.
}
