% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{binance_mytrades}
\alias{binance_mytrades}
\title{Get all trades on the Binance account}
\usage{
binance_mytrades(symbol, limit = 500, from_id)
}
\arguments{
\item{symbol}{string}

\item{limit}{number of trades to fetch}

\item{from_id}{optional trade id to fetch from}
}
\value{
data.table
}
\description{
Get all trades on the Binance account
}
\examples{
\dontrun{
binance_mytrades('ARKETH')
binance_mytrades(c('ARKBTC', 'ARKETH'))
}
}
