% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_open_interest.R
\name{binance_open_interest}
\alias{binance_open_interest}
\title{Binance Historical Open Interest}
\usage{
binance_open_interest(
  pair,
  api,
  interval,
  from,
  to,
  contract_type,
  quiet = FALSE
)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. Available options are:
\itemize{
\item \code{"fapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api/Open-Interest-Statistics}{fapi/v1/openInterestHist}. The ip weight is 0.
\item \code{"dapi"}: for endpoint \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api/Open-Interest-Statistics}{dapi/v1/openInterestHist}. The ip weight is 0.
}}

\item{interval}{Character. Time interval for open interest data. Default is \code{"1h"}. Available intervals are:
\itemize{
\item Minutely: \code{"5m"}, \code{"15m"} and \code{"30m"}.
\item Hourly: \code{"1h"}, \code{"2h"}, \code{"4h"}, \code{"6h"}, \code{"8h"} and \code{"12h"}.
\item Daily: \code{"1d"}.
}}

\item{from}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical data, only last 30 days are available.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()-lubridate::days(30)}.}

\item{to}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical data, only last 30 days are available.
If it is \code{missing}, the default, will be used as start date  \code{\link[=Sys.time]{Sys.time()}}.}

\item{contract_type}{Character. Used only if \code{api = "dapi"}. Available contract's types are:
\itemize{
\item \code{"all"}: for all types of contracts.
\item \code{"perpetual"}: for perpetual futures.
\item \code{"current_quarter"}: for futures with maturity in the current quarter.
\item \code{"next_quarter"}: for futures with maturity in the next quarter.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}} with 5 columns:
\itemize{
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, observation date.
\item \code{market}: Character, API.
\item \code{pair}: Character, trading pair.
\item \code{open_interest}: Numeric, open interest in base currency.
\item \code{open_interest_usd}: Numeric, open interest in USD.
}
}
\description{
Get the historical open interest for a trading pair. The data are only available for the last 30 days.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Historical open interest for BTCUSDT
binance_open_interest(pair = "BTCUSDT", 
                      api = "fapi", 
                      interval = "1d", 
                      from = Sys.time()-lubridate::days(15),
                      to = Sys.time()-lubridate::days(5))
                           

# Historical open interest for BTCUSD
binance_open_interest(pair = "BTCUSD", 
                      api = "dapi", 
                      interval = "1d", 
                      from = Sys.time()-lubridate::days(15),
                      to = NULL)
\dontshow{\}) # examplesIf}
}
\keyword{marketEndpoints}
