% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_open_interest_hist.R
\name{binance_open_interest_hist}
\alias{binance_open_interest_hist}
\title{Historical Open Interest Data}
\usage{
binance_open_interest_hist(pair,
                           api = "fapi", 
                           interval = "1d", 
                           from = NULL, 
                           to = NULL, 
                           contract_type = "all", 
                           quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"} or \code{"BTCUSD"}.}

\item{api}{Character, reference API. Available options are:
\itemize{
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#open-interest}{futures USD-m API}.
\item \code{"dapi"}: for \href{https://binance-docs.github.io/apidocs/delivery/en/#open-interest}{futures COIN-m API}.
}}

\item{interval}{Character. Default is \code{"1h"}. Time interval for open interest data. Available intervals are:
\itemize{
\item Minutely: \code{"5m"}, \code{"15m"} and \code{"30m"}.
\item Hourly: \code{"1h"}, \code{"2h"}, \code{"4h"}, \code{"6h"}, \code{"8h"} and \code{"12h"}.
\item Daily: \code{"1d"}.
}}

\item{from}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical data, only last 30 days are available.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()-lubridate::days(30)}.}

\item{to}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical data, only last 30 days are available.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()}.}

\item{contract_type}{Character. Used only if \code{api = "dapi"}. Available contract's types are:
\itemize{
\item \code{"all"}: for all types of contracts.
\item \code{"perpetual"}: for perpetual futures.
\item \code{"current_quarter"}: for futures with maturity in the current quarter.
\item \code{"next_quarter"}: for futures with maturity in the next quarter.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[=data.frame-class]{data.frame}} with 5 columns:
\itemize{
\item \code{date}: \code{"\link[=POSIXt-class]{POSIXt}"}, observation date.
\item \code{market}: Character, API.
\item \code{pair}: Character, trading pair.
\item \code{open_interest}: Numeric, open interest in base currency.
\item \code{open_interest_usd}: Numeric, open interest in USD.
}
}
\description{
Get the historical open interest data for a trading pair.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
The historical open interest data are only available for the last 30 days.
}
\examples{

# Historical open interest for BTCUSDT
binance_open_interest_hist(pair = "BTCUSDT", 
                           api = "fapi", 
                           interval = "1d", 
                           from = Sys.time()-lubridate::days(15),
                           to = Sys.time()-lubridate::days(5))

# Historical open interest for BTCUSD
binance_open_interest_hist(pair = "BTCUSD", 
                           api = "dapi", 
                           interval = "1d", 
                           from = Sys.time()-lubridate::days(15),
                           to = NULL)

}
