% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_open_orders}
\alias{binance_open_orders}
\title{Fetch open orders from the Binance account}
\usage{
binance_open_orders(symbol)
}
\arguments{
\item{symbol}{optional string}
}
\value{
data.table
}
\description{
Fetch open orders from the Binance account
}
\examples{
\dontrun{
binance_cancel_order('ARKETH', 8)
binance_cancel_order('ARKBTC', "myOrder7")
}
}
