% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_order_test}
\alias{binance_order_test}
\title{Manage orders on the Binance account}
\usage{
binance_order_test(symbol, side, type = "MARKET", quantity)
}
\arguments{
\item{symbol}{string}

\item{side}{enum}

\item{type}{enum}

\item{quantity}{number}
}
\value{
data.table
}
\description{
Manage orders on the Binance account
}
\examples{
\dontrun{
binance_mytrades('ARKETH')
binance_mytrades(c('ARKBTC', 'ARKETH'))
}
}
