% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ping.R
\name{binance_ping}
\alias{binance_ping}
\title{Ping to Binance REST API}
\usage{
binance_ping(api, quiet = FALSE)
}
\arguments{
\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://developers.binance.com/docs/binance-spot-api-docs/rest-api/general-endpoints#test-connectivity}{spot API}.
\item \code{"fapi"}: for \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/market-data/rest-api}{futures USD-m API}.
\item \code{"dapi"}: for \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/market-data/rest-api}{futures COIN-m API}.
\item \code{"eapi"}: for \href{https://developers.binance.com/docs/derivatives/option/market-data/Test-Connectivity}{options API}.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A logical value. It is \code{TRUE} if the connection was successful, otherwise it is \code{FALSE}.
}
\description{
Check the connection to the Binance API.
}
\details{
The IP weight for this API call is 1 for all the APIs.
}
\examples{
# Test connection to spot api
binance_ping("spot")
# Test connection to futures usd-m api
binance_ping("fapi")
# Test connection to futures coin-m api
binance_ping("dapi")
# Test connection to options api
binance_ping("eapi")

}
\keyword{generalEndpoints}
