% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_query.R
\name{binance_query}
\alias{binance_query}
\title{Query to Binance REST API}
\usage{
binance_query(
  api,
  path = NULL,
  query = list(),
  method = "GET",
  sign = FALSE,
  use_base_path = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://developers.binance.com/docs/binance-spot-api-docs/rest-api/general-api-information}{spot API}.
\item \code{"fapi"}: for \href{https://developers.binance.com/docs/derivatives/usds-margined-futures/general-info}{futures USD-m API}.
\item \code{"dapi"}: for \href{https://developers.binance.com/docs/derivatives/coin-margined-futures/general-info}{futures COIN-m API}.
\item \code{"eapi"}: for \href{https://developers.binance.com/docs/derivatives/option/general-info}{options API}.
\item \code{"sapi"}: for \href{https://binance-docs.github.io/apidocs/voptions/en/#wallet-endpoints}{wallet API}.
}}

\item{path}{Character vector. API path, \code{NULL} or \code{NA} elements will be excluded.}

\item{query}{Named list. Query parameters for the API call, \code{NULL} or \code{NA} elements will be excluded.}

\item{method}{Character. A method between \code{"\link[httr]{GET}"}, \code{"\link[httr]{POST}"} and \code{"\link[httr]{DELETE}"}.}

\item{sign}{Logical. Default is \code{FALSE}. \code{TRUE} if signature is required.}

\item{use_base_path}{Logical. When \code{TRUE}, the default, to \code{path} argument will be added a \code{base_bath} based on the selected API.
Available \code{base_bath} are:
\itemize{
\item \code{"spot"}: base path is \code{"api/v3"}.
\item \code{"fapi"}: base path is \code{"fapi/v1"}.
\item \code{"dapi"}: base path is \code{"dapi/v1"}.
\item \code{"eapi"}: base path is \code{"eapi/v1"}.
\item \code{"sapi"}: base path is \code{"sapi/v1"}.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
An R object parsed as character. See more on \code{\link[httr]{content}}
}
\description{
Execute a query to Binance REST API.
}
\examples{
# Execute a call to spot API with base path
binance_query(api = "spot", 
              path = "time", 
              query = NULL, 
              use_base_path = TRUE)

# Execute a call to spot API without base path
binance_query(api = "spot", 
              path = c("api", "v3", "time"), 
              query = NULL,
              use_base_path = FALSE)

}
\keyword{utils}
