% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_query_order}
\alias{binance_query_order}
\title{Query order on the Binance account}
\usage{
binance_query_order(symbol, order_id, client_order_id)
}
\arguments{
\item{symbol}{string}

\item{order_id}{optional number}

\item{client_order_id}{optional string}
}
\value{
data.table
}
\description{
Query order on the Binance account
}
\examples{
\dontrun{
binance_query_order('ARKETH', 8)
binance_query_order('ARKBTC', 'myOrder7')
}
}
