% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_secret}
\alias{binance_secret}
\title{Look up Binance API secret stored in the environment}
\usage{
binance_secret()
}
\value{
string
}
\description{
Look up Binance API secret stored in the environment
}
\keyword{internal}
