% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_sign}
\alias{binance_sign}
\title{Sign the query string for Binance}
\usage{
binance_sign(params, time = timestamp())
}
\arguments{
\item{params}{list}

\item{time}{optional string}
}
\value{
list
}
\description{
Sign the query string for Binance
}
\examples{
\dontrun{
binance_sign(list(foo = 'bar', z = 4))
}
}
\keyword{internal}
