% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_symbols}
\alias{binance_symbols}
\title{Get all currently valid symbol names from Binance}
\usage{
binance_symbols()
}
\value{
character vector
}
\description{
Get all currently valid symbol names from Binance
}
