% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ticker24h.R
\name{binance_ticker24h}
\alias{binance_ticker24h}
\title{Binance 24-Hour Ticker Statistics}
\usage{
binance_ticker24h(pair, api, type, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}. Multiple pairs are allowed only if \code{api} is \code{"spot"}.
If \code{missing}, the default, will be returned the 24hr ticker for all the trading pairs.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for endpoint \href{https://binance-docs.github.io/apidocs/spot/en/#24hr-ticker-price-change-statistics}{api/v3/ticker/24hr}.
The ip weight depends on the number of pairs requested. The maximum ip weight is 80.
\item \code{"fapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/futures/en/#24hr-ticker-price-change-statistics}{fapi/v1/ticker/24hr}. The ip weight is 1.
\item \code{"dapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/delivery/en/#24hr-ticker-price-change-statistics}{dapi/v1/ticker/24hr}. The ip weight is 1.
\item \code{"eapi"}: for endpoint \href{https://binance-docs.github.io/apidocs/voptions/en/#24hr-ticker-price-change-statistics}{eapi/v1/ticker}. The ip weight is 5.
}}

\item{type}{Character. Type of ticker data. Used only if \code{api = "spot"}. Default is \code{"full"}. Available options are:
\itemize{
\item \code{"mini"}: data without ask and bid prices and quantities.
\item \code{"full"}: complete ticker data.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}} with 13 columns containing 24-hour ticker statistics, including:
open, high, low, close prices, volume, and more.
}
\description{
Get 24-hour ticker statistics for a specified trading pair from the selected reference API.
}
\examples{
# Get full 24-hour ticker for all pairs
binance_ticker24h(api = "spot")
binance_ticker24h(api = "fapi")
binance_ticker24h(api = "dapi")
binance_ticker24h(api = "eapi")

# Get full 24-hour ticker for BTCUSDT
binance_ticker24h(pair = "BTCUSDT", api = "spot", type = "full")

# Get mini 24-hour ticker for BTCUSDT
binance_ticker24h(pair = "BTCUSDT", api = "spot", type = "mini")

# Get 24-hour ticker for BTCUSDT 
binance_ticker24h(pair = "BTCUSDT", api = "fapi")

# Get 24-hour ticker for BTCUSD_PERP
binance_ticker24h(pair = "BTCUSD_PERP", api = "dapi")

# Get 24-hour ticker for a put option on BTCUSDT
binance_ticker24h(pair = "BTC-240628-30000-P", api = "eapi")

}
\keyword{marketEndpoints}
