% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_ticker_all_books}
\alias{binance_ticker_all_books}
\title{Get latest Binance bids and asks on all symbol pairs}
\usage{
binance_ticker_all_books()
}
\value{
\code{data.table}
}
\description{
Get latest Binance bids and asks on all symbol pairs
}
