% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticket.R
\name{binance_ticker_all_prices}
\alias{binance_ticker_all_prices}
\title{Get latest prices on all symbols}
\usage{
binance_ticker_all_prices()
}
\value{
data.table
}
\description{
Get latest prices on all symbols
}
