% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_ticker_book}
\alias{binance_ticker_book}
\title{Get last bids and asks for a symbol or all symbols}
\usage{
binance_ticker_book(symbol)
}
\arguments{
\item{symbol}{optional string}
}
\value{
data.table
}
\description{
Get last bids and asks for a symbol or all symbols
}
