% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ticker_day.R
\name{binance_ticker_day}
\alias{binance_ticker_day}
\title{Symbol Trading Day Ticker}
\usage{
binance_ticker_day(pair, time_zone = 0, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}. Multiple pairs are allowed with a maximum of 100 per request.}

\item{time_zone}{Character, reference time zone.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[tibble]{tibble}} with 14 columns:
\itemize{
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, open date of the reference period.
\item \code{date_close}: \code{\link[=POSIXt-class]{POSIXt}}, close date of the reference period.
\item \code{symbol}: Character, trading pair.
\item \code{market}: Character, reference API.
\item \code{first_id}: Numeric, first trade id.
\item \code{last_id}: Numeric, last trade id.
\item \code{open}: Numeric, open price.
\item \code{high}: Numeric, highest price.
\item \code{low}: Numeric, lowest price.
\item \code{last}: Numeric, last price.
\item \code{weighted_price}: Numeric, volume-weighted price.
\item \code{volume}: Numeric, volume.
\item \code{volume_quote}: Numeric, volume in terms of quote asset.
\item \code{trades}: Numeric, number of trades
}
}
\description{
Price change statistics for a trading day.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Get day ticker for BTCUSDT
binance_ticker_day("BTCUSDT")
\dontshow{\}) # examplesIf}
}
\keyword{marketEndpoints}
