% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_vision_klines.R
\name{binance_vision_klines}
\alias{binance_vision_klines}
\title{binance_vision_klines}
\usage{
binance_vision_klines(
  api = "spot",
  pair = NULL,
  interval = "1d",
  from = NULL,
  to = NULL,
  quiet = FALSE
)
}
\arguments{
\item{pair}{a character containing the Pair of interest, for example \code{BTCUSDT}.}

\item{interval}{character object containing the time interval, for example \verb{1m} for 1 minute data, \verb{1h} for 1 hour data.}

\item{quiet}{logical, if TRUE some messages will be displayed.}

\item{day_date}{a character or a Date object containing the date of witch we would like to obtain the data. Note
that the data available refer to the day before. For more recent data is better to use the REST api or the WebSocket.}
}
\value{
Returns a \code{tibble} object
}
\description{
Historical Kline/Candlestick Data for a pair from the database.
}
\examples{

}
