% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws_cleaner.R
\name{binance_ws_cleaner}
\alias{binance_ws_cleaner}
\title{Cleaner binance websocket}
\usage{
binance_ws_cleaner(data)
}
\arguments{
\item{data}{Tibble}
}
\description{
Clean and structure the different websocket responses.
}
