% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws.R
\name{binance_ws_socket}
\alias{binance_ws_socket}
\title{binance_spot_socket}
\usage{
binance_ws_socket(
  pair = "BTCUSDT",
  api = "spot",
  subscription = "kline",
  interval = "1m",
  contract_type = NULL,
  update_speed = 1000,
  nobs_max = NULL,
  quiet = FALSE
)
}
\arguments{
\item{pair}{Character, the trading pair of interest, e.g., "BTCUSDT".}

\item{api}{Character, specifying the reference API. Available options are \code{spot}, \code{fapi}, \code{dapi}, \code{eapi}.}

\item{subscription}{Character}

\item{interval}{Character, the time interval for Klines data.
Acceptable intervals are "1s", "1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d", "3d", "1w", and "1M".}

\item{contract_type}{Character, only used if \code{api = "dapi"}. Choose from:
\itemize{
\item "perpetual": For perpetual futures.
\item "current_quarter": For futures with a maturity in the current quarter.
\item "next_quarter": For futures with a maturity in the next quarter.
}}

\item{update_speed}{time in milliseconds}

\item{nobs_max}{Integer, maximum number of observations to save.}

\item{quiet}{Logical, suppress informational messages if TRUE.}
}
\description{
api websocket
}
