% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binancer.R
\docType{package}
\name{binancer-package}
\alias{binancer}
\alias{binancer-package}
\title{binancer: Binance cryptocurrency data tools}
\description{
Tools for retrieving cryptocurrency market data from Binance REST endpoints,
Binance Vision public archive files, and Binance WebSocket streams. Authenticated
account and order helpers require credentials supplied by the user with
\code{\link[=binance_credentials]{binance_credentials()}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beniamino98/binancer}
  \item Report bugs at \url{https://github.com/beniamino98/binancer/issues}
}

}
\author{
\strong{Maintainer}: Beniamino Sartini \email{beniaminosartini@gmail.com}

}
\keyword{internal}
