% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binancer.R
\docType{package}
\name{binancer}
\alias{binancer-package}
\alias{binancer}
\title{binancer: Wrapper for Binance REST and Websocket API in R.}
\description{
Wrapper for Binance REST API in R
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.cryptoverser.org}
  \item Report bugs at \url{https://github.com/beniamino98/binancer/issues}
}

}
