% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binanceWebSocket.R
\name{binanceWebSocket}
\alias{binanceWebSocket}
\title{R6 Class for Binance Web Socket}
\value{
An R6 generator for \code{binanceWebSocket} objects.
}
\description{
Binance web socket description
}
\details{
Creates a WebSocket client for Binance stream subscriptions.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
ws <- binanceWebSocket$new(
  pair = "BTCUSDT",
  api = "spot",
  subscription = "kline",
  interval = "1m"
)
ws$connect()
ws$close()
\dontshow{\}) # examplesIf}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{stream}}{Get all streams}

\item{\code{info}}{Get all streams info}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-binanceWebSocket-new}{\code{binanceWebSocket$new()}}
\item \href{#method-binanceWebSocket-add_depth_snapshot}{\code{binanceWebSocket$add_depth_snapshot()}}
\item \href{#method-binanceWebSocket-add_trades_snapshot}{\code{binanceWebSocket$add_trades_snapshot()}}
\item \href{#method-binanceWebSocket-add_klines_snapshot}{\code{binanceWebSocket$add_klines_snapshot()}}
\item \href{#method-binanceWebSocket-subscribe}{\code{binanceWebSocket$subscribe()}}
\item \href{#method-binanceWebSocket-unsubscribe}{\code{binanceWebSocket$unsubscribe()}}
\item \href{#method-binanceWebSocket-connect}{\code{binanceWebSocket$connect()}}
\item \href{#method-binanceWebSocket-close}{\code{binanceWebSocket$close()}}
\item \href{#method-binanceWebSocket-get_stream}{\code{binanceWebSocket$get_stream()}}
\item \href{#method-binanceWebSocket-get_info}{\code{binanceWebSocket$get_info()}}
\item \href{#method-binanceWebSocket-clone}{\code{binanceWebSocket$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-new"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{binanceWebSocket} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$new(
  pair = "BTCUSDT",
  api = "spot",
  subscription,
  interval,
  update_speed
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{\code{api}}{Character. Reference API. Available options are \code{"spot"} and \code{"fapi"}.}

\item{\code{subscription}}{type of subscription, can be \code{aggTrade},\code{bookTicker},\code{depth},\code{kline},\code{miniTicker},\code{ticker},\code{trade}.}

\item{\code{interval}}{time interval, can be \verb{1s},\verb{1m},\verb{3m},\verb{5m},\verb{15m},\verb{30m},\verb{1h},\verb{2h},\verb{4h},\verb{6h},\verb{12h},\verb{1d},\verb{3d},\verb{1w},\verb{1M}.}

\item{\code{update_speed}}{update speed in millisecond, for \code{spot} api \verb{1000ms}. For others apis \verb{500ms}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{binanceWebSocket} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-add_depth_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-add_depth_snapshot}{}}}
\subsection{Method \code{add_depth_snapshot()}}{
Add a snapshot of the order book.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$add_depth_snapshot(pair, subscription, update_speed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{\code{subscription}}{type of subscription, \code{depth}.}

\item{\code{update_speed}}{update speed in millisecond, for \code{spot} api \verb{1000ms}. For others apis \verb{500ms}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-add_trades_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-add_trades_snapshot}{}}}
\subsection{Method \code{add_trades_snapshot()}}{
Add trades
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$add_trades_snapshot(pair, subscription, from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{\code{subscription}}{type of subscription, \code{aggTrades} or \code{trades}.}

\item{\code{from}}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical data.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()-lubridate::days(1)}.}

\item{\code{to}}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical data.
If it is \code{missing}, the default, will be used as end date \code{\link[=Sys.time]{Sys.time()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-add_klines_snapshot"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-add_klines_snapshot}{}}}
\subsection{Method \code{add_klines_snapshot()}}{
Add a snapshot of the klines.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$add_klines_snapshot(pair, subscription, interval, from, to)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{\code{subscription}}{type of subscription, \code{klines}.}

\item{\code{interval}}{time interval, can be \verb{1s},\verb{1m},\verb{3m},\verb{5m},\verb{15m},\verb{30m},\verb{1h},\verb{2h},\verb{4h},\verb{6h},\verb{12h},\verb{1d},\verb{3d},\verb{1w},\verb{1M}.}

\item{\code{from}}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Start time for historical data.
If it is \code{missing}, the default, will be used as start date \code{Sys.time()-lubridate::days(1)}.}

\item{\code{to}}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. End time for historical data.
If it is \code{missing}, the default, will be used as end date \code{\link[=Sys.time]{Sys.time()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-subscribe}{}}}
\subsection{Method \code{subscribe()}}{
Subscribe to a stream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$subscribe(pair, subscription, interval, update_speed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{\code{subscription}}{type of subscription, can be \code{aggTrade},\code{bookTicker},\code{depth},\code{kline},\code{miniTicker},\code{ticker},\code{trade}.}

\item{\code{interval}}{time interval, can be \verb{1s},\verb{1m},\verb{3m},\verb{5m},\verb{15m},\verb{30m},\verb{1h},\verb{2h},\verb{4h},\verb{6h},\verb{12h},\verb{1d},\verb{3d},\verb{1w},\verb{1M}.}

\item{\code{update_speed}}{update speed in millisecond, for \code{spot} api \verb{1000ms}. For others apis \verb{500ms}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-unsubscribe"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-unsubscribe}{}}}
\subsection{Method \code{unsubscribe()}}{
Unsubscribe from a stream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$unsubscribe(pair, subscription, interval, update_speed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{\code{subscription}}{type of subscription, can be \code{aggTrade},\code{bookTicker},\code{depth},\code{kline},\code{miniTicker},\code{ticker},\code{trade}.}

\item{\code{interval}}{time interval, can be \verb{1s},\verb{1m},\verb{3m},\verb{5m},\verb{15m},\verb{30m},\verb{1h},\verb{2h},\verb{4h},\verb{6h},\verb{12h},\verb{1d},\verb{3d},\verb{1w},\verb{1M}.}

\item{\code{update_speed}}{update speed in millisecond, for \code{spot} api \verb{1000ms}. For others apis \verb{500ms}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-connect"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-connect}{}}}
\subsection{Method \code{connect()}}{
Open the connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-close"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-close}{}}}
\subsection{Method \code{close()}}{
Close all the connections.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-get_stream"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-get_stream}{}}}
\subsection{Method \code{get_stream()}}{
Get a stream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$get_stream(pair, subscription, interval, update_speed, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{type of subscription, can be \code{aggTrade},\code{bookTicker},\code{depth},\code{kline},\code{miniTicker},\code{ticker},\code{trade}.}

\item{\code{interval}}{time interval, can be \verb{1s},\verb{1m},\verb{3m},\verb{5m},\verb{15m},\verb{30m},\verb{1h},\verb{2h},\verb{4h},\verb{6h},\verb{12h},\verb{1d},\verb{3d},\verb{1w},\verb{1M}.}

\item{\code{update_speed}}{update speed in millisecond, for \code{spot} api \verb{1000ms}. For others apis \verb{500ms}.}

\item{\code{id}}{Numeric stream id. If supplied, \code{pair}, \code{subscription}, \code{interval}, and \code{update_speed} are ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-get_info"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-get_info}{}}}
\subsection{Method \code{get_info()}}{
Get the info of all the streams opened.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$get_info(pair, subscription, interval, update_speed, id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pair}}{pair}

\item{\code{subscription}}{type of subscription, can be \code{aggTrade},\code{bookTicker},\code{depth},\code{kline},\code{miniTicker},\code{ticker},\code{trade}.}

\item{\code{interval}}{time interval, can be \verb{1s},\verb{1m},\verb{3m},\verb{5m},\verb{15m},\verb{30m},\verb{1h},\verb{2h},\verb{4h},\verb{6h},\verb{12h},\verb{1d},\verb{3d},\verb{1w},\verb{1M}.}

\item{\code{update_speed}}{update speed in millisecond, for \code{spot} api \verb{1000ms}. For others apis \verb{500ms}.}

\item{\code{id}}{Numeric stream id. If supplied, \code{pair}, \code{subscription}, \code{interval}, and \code{update_speed} are ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binanceWebSocket-clone"></a>}}
\if{latex}{\out{\hypertarget{method-binanceWebSocket-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{binanceWebSocket$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
