% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{close_long_limit}
\alias{close_long_limit}
\title{Close long position at limited price}
\usage{
close_long_limit(symbol, price, quantity, time_in_force = "GTC")
}
\arguments{
\item{symbol}{string}

\item{price}{numeric}

\item{quantity}{numeric}

\item{time_in_force}{enum}
}
\value{
data.table
}
\description{
Close long position at limited price
}
