% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{close_long_market}
\alias{close_long_market}
\title{Close long position at market price}
\usage{
close_long_market(symbol, quantity)
}
\arguments{
\item{symbol}{string}

\item{quantity}{numeric}
}
\value{
data.table
}
\description{
Close long position at market price
}
