% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_prep.R
\name{coin_data_importer}
\alias{coin_data_importer}
\title{Fetch historical coin data}
\usage{
coin_data_importer<-function(symbol, interval = interval, start_time = start_time, end_time = end_time)
}
\arguments{
\item{symbol}{The binance market of interest (e.g. the Ethereum-Bitcoin market "ETHBTC")}

\item{interval}{The interval at which historical data is pulled ("1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d", "3d", "1w", "1M")}

\item{start_time}{The starting date of the historical data}

\item{end_time}{The end date of the historical data}
}
\description{
This function serves to pull historical market data for a given symbol
}
