% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_utils.R
\name{coindata_firstloop}
\alias{coindata_firstloop}
\title{first round of coin dataset in backtest}
\usage{
coindata_firstloop(
  data,
  portval = portval,
  min_momentum_count = min_momentum_count,
  min_trade = min_trade,
  timeperiod = timeperiod,
  i = i
)
}
\arguments{
\item{fees}{}
}
\description{
first round of coin dataset in backtest
}
