% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{convert_position_risks}
\alias{convert_position_risks}
\title{Convert columns of positions into numeric.}
\usage{
convert_position_risks(positions)
}
\arguments{
\item{positions}{data.table}
}
\value{
data.table
}
\description{
Convert columns of positions into numeric.
}
