% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{execute_usdm_order}
\alias{execute_usdm_order}
\title{Execute an order on USDM}
\usage{
execute_usdm_order(order, ...)
}
\arguments{
\item{order}{string}

\item{...}{list}
}
\value{
data.table
}
\description{
Execute an order on USDM
}
