% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{generate_indicator}
\alias{generate_indicator}
\title{Generate trade indicator information}
\usage{
generate_indicator(
  historical_data = historical_data,
  indicator = settings$indicator
)
}
\value{
historical_data
}
\description{
Generate trade indicator information
}
