% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klines.R
\name{get_klines}
\alias{get_klines}
\title{Get kline/candlestick data}
\usage{
get_klines(symbol, interval, limit)
}
\arguments{
\item{symbol}{string}

\item{interval}{enum}

\item{limit}{int}
}
\value{
data.table
}
\description{
Get kline/candlestick data
}
