% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{historical}
\alias{historical}
\title{Fetch historical coin data}
\usage{
historical(symbol, interval = "1d", start_time = "2021-01-01", end_time = NULL)
}
\arguments{
\item{interval}{The interval at which historical data is pulled ("1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d", "3d", "1w", "1M")}

\item{start_time}{The starting date of the historical data}

\item{end_time}{The end date of the historical data}

\item{base_currency}{The base currency being used for trading (e.g. "BTC", "ETC")}

\item{symbol}{The binance market of interest (e.g. the Ethereum-Bitcoin market "ETHBTC")}
}
\value{
historical_data
}
\description{
This function serves to pull all historical market data that
}
