% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huobi.R
\name{huobi_klines}
\alias{huobi_klines}
\title{Get kline/candlestick data from Huobi}
\usage{
huobi_klines(symbol, period = c("1min", "5min", "15min", "30min",
  "60min", "1day", "1mon", "1week", "1year"), size = 500)
}
\arguments{
\item{symbol}{string}

\item{period}{enum}

\item{size}{int}
}
\value{
list where you are looking for the \code{data} element
}
\description{
Get kline/candlestick data from Huobi
}
