% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huobi.R
\name{huobi_query}
\alias{huobi_query}
\title{Request the Huobi API}
\usage{
huobi_query(endpoint, method = "GET", params = list(), sign = FALSE,
  retry = method == "GET")
}
\arguments{
\item{endpoint}{string}

\item{method}{HTTP request method}

\item{params}{list}

\item{sign}{if signature required}

\item{retry}{allow retrying the query on failure}
}
\value{
R object
}
\description{
Request the Huobi API
}
\keyword{internal}
