% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{long_stop_market_all}
\alias{long_stop_market_all}
\title{Stop all loss from a long position at market price}
\usage{
long_stop_market_all(symbol, stop_price)
}
\arguments{
\item{symbol}{string}

\item{stop_price}{numeric}
}
\value{
data.table
}
\description{
Stop all loss from a long position at market price
}
