% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{open_long_limit}
\alias{open_long_limit}
\title{Open long position at a limited price}
\usage{
open_long_limit(symbol, price, quantity, time_in_force = "GTC")
}
\arguments{
\item{symbol}{string}

\item{price}{numeric}

\item{quantity}{numeric}

\item{time_in_force}{enum}
}
\value{
data.table
}
\description{
Open long position at a limited price
}
