% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{open_long_market}
\alias{open_long_market}
\title{Open long position at market price}
\usage{
open_long_market(symbol, quantity)
}
\arguments{
\item{symbol}{string}

\item{quantity}{numeric}
}
\value{
data.table
}
\description{
Open long position at market price
}
