% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{open_short_limit}
\alias{open_short_limit}
\title{Create an open short limit order object}
\usage{
open_short_limit(symbol, price, quantity, time_in_force = "GTC")
}
\arguments{
\item{symbol}{string}

\item{price}{numeric}

\item{quantity}{numeric}

\item{time_in_force}{enum}
}
\value{
data.table
}
\description{
Create an open short limit order object
}
