% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{open_short_market}
\alias{open_short_market}
\title{Open short position at market price}
\usage{
open_short_market(symbol, quantity)
}
\arguments{
\item{symbol}{string}

\item{quantity}{numeric}
}
\value{
data.table
}
\description{
Open short position at market price
}
