% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_prep.R
\docType{data}
\name{settings}
\alias{settings}
\title{Set backtest settings}
\format{
An object of class \code{environment} of length 0.
}
\usage{
settings
}
\arguments{
\item{min_trade}{}
}
\description{
Set backtest settings
}
\keyword{datasets}
