% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{signature}
\alias{signature}
\title{Sign the query string for Binance}
\usage{
signature(params)
}
\arguments{
\item{params}{list}
}
\value{
string
}
\description{
Sign the query string for Binance
}
\examples{
\dontrun{
signature(list(foo = 'bar', z = 4))
}
}
\keyword{internal}
