% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timestamp}
\alias{timestamp}
\title{Return current UNIX timestamp in millisecond}
\usage{
timestamp()
}
\value{
milliseconds since Jan 1, 1970
}
\description{
Return current UNIX timestamp in millisecond
}
\keyword{internal}
