% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_cancel_order_by_client_order_id}
\alias{usdm_v1_cancel_order_by_client_order_id}
\title{Cancel an active order by its client order id.}
\usage{
usdm_v1_cancel_order_by_client_order_id(symbol, client_order_id)
}
\arguments{
\item{symbol}{string}

\item{client_order_id}{string}
}
\value{
data.table
}
\description{
Cancel an active order by its client order id.
}
