% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_cancel_order_by_id}
\alias{usdm_v1_cancel_order_by_id}
\title{Cancel an active order by its order id.}
\usage{
usdm_v1_cancel_order_by_id(symbol, order_id)
}
\arguments{
\item{symbol}{string}

\item{order_id}{integer}
}
\value{
data.table
}
\description{
Cancel an active order by its order id.
}
