% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_change_initial_leverage}
\alias{usdm_v1_change_initial_leverage}
\title{Change user's initial leverage of specific symbol market.}
\usage{
usdm_v1_change_initial_leverage(symbol, leverage)
}
\arguments{
\item{symbol}{string}

\item{leverage}{integer}
}
\value{
list
}
\description{
Change user's initial leverage of specific symbol market.
}
