% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_change_margin_type}
\alias{usdm_v1_change_margin_type}
\title{Change the margin type of specific symbol market.}
\usage{
usdm_v1_change_margin_type(symbol, type = c("ISOLATED", "CROSSED"))
}
\arguments{
\item{symbol}{string}

\item{type}{enum}
}
\value{
list
}
\description{
Change the margin type of specific symbol market.
}
