% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_filters}
\alias{usdm_v1_filters}
\title{Get the filters of a symbol from Binance USDM}
\usage{
usdm_v1_filters(symbol, info = usdm_v1_exchange_info())
}
\arguments{
\item{symbol}{string}

\item{info}{list}
}
\value{
\code{data.table}
}
\description{
Get the filters of a symbol from Binance USDM
}
