% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_new_order}
\alias{usdm_v1_new_order}
\title{Open new order on the Binance USDM account}
\usage{
usdm_v1_new_order(
  symbol,
  side = BINANCE$SIDE,
  position_side = BINANCE$USDM$POSITION_SIDE,
  type = BINANCE$USDM$TYPE,
  ...
)
}
\arguments{
\item{symbol}{string}

\item{side}{enum}

\item{position_side}{enum}

\item{type}{enum}

\item{...}{list}
}
\value{
data.table
}
\description{
This function serves as a low level entry for order classes.
Do not use it directly.
}
