% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_open_orders}
\alias{usdm_v1_open_orders}
\title{Get all open orders of a symbol on USDM.}
\usage{
usdm_v1_open_orders(symbol)
}
\arguments{
\item{symbol}{optional string}
}
\value{
data.table
}
\description{
Get all open orders on a symbol. Careful when accessing this with no symbol.
Weight: 1 for a single symbol; 40 when the symbol parameter is omitted.
}
