% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_premium_index}
\alias{usdm_v1_premium_index}
\title{Get mark/index price for a symbol}
\usage{
usdm_v1_premium_index(symbol)
}
\arguments{
\item{symbol}{string}
}
\value{
\code{data.table}
}
\description{
Get mark/index price for a symbol
}
