% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v2_position_risks}
\alias{usdm_v2_position_risks}
\title{Get positions of a symbol or all symbols on USDM.}
\usage{
usdm_v2_position_risks(symbol)
}
\arguments{
\item{symbol}{optional string}
}
\value{
data.table
}
\description{
Get positions of a symbol or all symbols on USDM.
}
