% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{version_check}
\alias{version_check}
\title{Look up Binance API secret stored in the environment}
\usage{
version_check()
}
\value{
string
}
\description{
Look up Binance API secret stored in the environment
}
\keyword{internal}
