% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{binance_account}
\alias{binance_account}
\title{Get current account information, including balances}
\usage{
binance_account()
}
\value{
data.table
}
\description{
Get current account information, including balances
}
