% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_account.R
\name{binance_account_info}
\alias{binance_account_info}
\title{Binance Account Information}
\usage{
binance_account_info()
}
\value{
A \code{\link[tibble]{tibble}}.
}
\description{
Get current general Binance account information.
}
\keyword{AccountEndpoints}
