% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_account.R
\name{binance_account_trades}
\alias{binance_account_trades}
\title{Get trades informations}
\usage{
binance_account_trades(symbol, from, to, from_id, limit = 1000)
}
\value{
data.table
}
\description{
Get trades informations
}
