% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_api.R
\name{binance_api}
\alias{binance_api}
\title{Binance REST API Call Function}
\usage{
binance_api(
  api = NULL,
  path = NULL,
  query = NULL,
  use_base_path = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{api}{Character, specifying the reference API. Available options include:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#introduction}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#introduction}{Futures USD-M API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#introduction}{Futures Coin-M API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#introduction}{Options API}.
}}

\item{path}{Character vector, specifying the API path. This vector represents the structure of the API endpoint.}

\item{query}{Named list, containing query parameters for the API call.}

\item{use_base_path}{Logical, if \code{TRUE}, the appropriate base path will be automatically added to the \code{path} argument based on the selected API type. The base paths are as follows:
\itemize{
\item "spot": For the Spot API, the base path is "api/v3".
\item "fapi": For the Futures USD-M API, the base path is "fapi/v1".
\item "dapi": For the Futures Coin-M API, the base path is "dapi/v1".
\item "eapi": For the Options API, the base path is "eapi/v1".
}}

\item{quiet}{Logical, suppress informational messages if TRUE.}
}
\description{
Execute an API call to the Binance REST API, allowing access to various endpoints across different Binance API types.
}
\examples{

# Example: Execute an API call to "api/v3/time" with automatic base path addition (use_base_path = TRUE) for the Spot API.
binance_api(api = "spot", path = c("time"), query = NULL)

# Example: Execute an API call to "api/v3/time" without automatic base path addition (use_base_path = FALSE) for the Spot API.
binance_api(api = "spot", path = c("api", "v3", "time"), use_base_path = FALSE, query = NULL)

}
