% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_avg_price}
\alias{binance_avg_price}
\title{Get current average price for a symbol}
\usage{
binance_avg_price(symbol)
}
\arguments{
\item{symbol}{string}
}
\value{
\code{data.table}
}
\description{
Get current average price for a symbol
}
