% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_balances}
\alias{binance_balances}
\title{Get current Binance balances in a nice table}
\usage{
binance_balances(threshold = -1, usdt = FALSE)
}
\arguments{
\item{threshold}{show assets with greater number of coins}

\item{usdt}{to include balance in USDT too}
}
\value{
data.table
}
\description{
Get current Binance balances in a nice table
}
