% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_cancel_order}
\alias{binance_cancel_order}
\title{Cancel order on the Binance account}
\usage{
binance_cancel_order(symbol, order_id, client_order_id)
}
\arguments{
\item{symbol}{string}

\item{order_id}{optional number}

\item{client_order_id}{optional string}
}
\value{
data.table
}
\description{
Cancel order on the Binance account
}
\examples{
\dontrun{
binance_cancel_order('ARKETH')
binance_cancel_order('ARKBTC', client_order_id = 'myOrder7')
}
}
