% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_check_credentials}
\alias{binance_check_credentials}
\title{Check if Binance credentials were set previously}
\usage{
binance_check_credentials()
}
\value{
No return values, but fails when credentials were not set.
}
\description{
Check if Binance credentials were set previously
}
\keyword{internal}
