% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_credentials.R
\name{binance_credentials}
\alias{binance_credentials}
\title{Set Binance API Keys}
\usage{
binance_credentials(key, secret)
}
\arguments{
\item{key}{Character. Api key.}

\item{secret}{Character. Api secret.}
}
\value{
No return values, setting config in the package namespace.
}
\description{
Sets the API key and secret to interact with the Binance API
}
\examples{
\dontrun{
# Add api keys 
binance_credentials('foo', 'bar')
# Remove api keys 
binance_credentials()
}

}
\keyword{apikey}
