% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binance_credentials}
\alias{binance_credentials}
\title{Sets the API key and secret to interact with the Binance API}
\usage{
binance_credentials(key, secret)
}
\arguments{
\item{key}{string}

\item{secret}{string}
}
\description{
Sets the API key and secret to interact with the Binance API
}
\examples{
\dontrun{
binance_credentials('foo', 'bar')
}
}
