% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_depth.R
\name{binance_depth}
\alias{binance_depth}
\title{Binance Order Book Snapshot}
\usage{
binance_depth(pair, api, quiet = FALSE)
}
\arguments{
\item{pair}{Character, trading pair, e.g. "BTCUSDT" or "BTCUSD_PERP" for COIN-m API.}

\item{api}{Character, reference API. Available options are:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#order-book}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#order-book}{Futures USD-m API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#order-book}{Futures COIN-m API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#order-book}{Options API}.
}}

\item{quiet}{Logical, if \code{TRUE} suppress informational and warnings. Default is \code{FALSE}.}
}
\value{
A tibble (data frame) with 6 columns:
\itemize{
\item \code{date}: Datetime, the time of the order book snapshot.
\item \code{market}: Character, reference API.
\item \code{pair}: Character, selected trading pair.
\item \code{side}: Character, side of the order. "ASK" for sell orders and "BID" for buy orders.
\item \code{price}: Numeric, price levels in the order book.
\item \code{quantity}: Numeric, quantity correspondent to each price level in the order book.
}
}
\description{
Retrieve a snapshot of the order book for a specified trading pair.
}
\examples{

# Get the order book for BTCUSDT in Spot market.
binance_depth(pair = "BTCUSDT", api = "spot")

# Get the order book for BTCUSDT in Futures USD-M market.
binance_depth(pair = "BTCUSDT", api = "fapi")

# Get the order book for BTCUSD_PERP in Futures COIN-M market.
binance_depth(pair = "BTCUSD_PERP", api = "dapi")

# Get the order book for a Put option on BTCUSDT 
binance_depth(pair = "BTC-240628-30000-P", api = "eapi")

}
