% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_depth.R
\name{binance_depth}
\alias{binance_depth}
\title{Binance Order Book Snapshot}
\usage{
binance_depth(pair, 
              api, 
              quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://binance-docs.github.io/apidocs/spot/en/#order-book}{spot API}.
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#order-book}{futures USD-m API}.
\item \code{"dapi"}: for \href{https://binance-docs.github.io/apidocs/delivery/en/#order-book}{futures COIN-m API}.
\item \code{"eapi"}: for \href{https://binance-docs.github.io/apidocs/voptions/en/#order-book}{options API}.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[=data.frame-class]{data.frame}} with 7 columns:
\itemize{
\item \code{last_update_id}: Integer, id of the last snapshot.
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, time of the snapshot.
\item \code{market}: Character, reference API.
\item \code{pair}: Character, trading pair.
\item \code{side}: Character, side of the order. Can be \code{"ASK"} or \code{"BID"}.
\item \code{price}: Numeric, price levels.
\item \code{quantity}: Numeric, quantity in each price level.
}
}
\description{
Retrieve a snapshot of the order book for a specified trading pair.
}
\examples{

# Get the order book for BTCUSDT
binance_depth(pair = "BTCUSDT", api = "spot")
binance_depth(pair = "BTCUSDT", api = "fapi")

# Get the order book for BTCUSD_PERP
binance_depth(pair = "BTCUSD_PERP", api = "dapi")

# Get the order book for a put option on BTC 
binance_depth(pair = "BTC-240628-30000-P", api = "eapi")

}
