% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_depth}
\alias{binance_depth}
\title{Get orderbook depth data from Binance}
\usage{
binance_depth(symbol, limit = c(5, 10, 20, 50, 100, 500, 1000))
}
\arguments{
\item{symbol}{string}

\item{limit}{int}
}
\value{
data.table
}
\description{
Get orderbook depth data from Binance
}
\examples{
\dontrun{
binance_depth('ETHUSDT')
binance_depth('ETHUSDT', limit = 1000)
}
}
