% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_depth}
\alias{binance_depth}
\title{Get orderbook depth data from Binance}
\usage{
binance_depth(symbol, limit)
}
\arguments{
\item{symbol}{string}

\item{limit}{int optional}
}
\value{
\code{data.table}
}
\description{
Get orderbook depth data from Binance
}
\examples{
\dontrun{
binance_depth('ETHUSDT')
binance_depth('ETHUSDT', limit = 1000)
}
}
