% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_dust_conversion}
\alias{binance_dust_conversion}
\title{Convert 'dust' assets in the Binance account into BNB}
\usage{
binance_dust_conversion(dust)
}
\arguments{
\item{asset}{The ticker of the asset you wish to convert to BNB (e.g 'ETH')}
}
\value{
data.table
}
\description{
Convert 'dust' assets in the Binance account into BNB
}
