% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_klines.R
\name{binance_klines}
\alias{binance_klines}
\title{Get Klines Data from Binance}
\usage{
binance_klines(pair,
               api = "spot",
               interval, 
               from, 
               to,
               contract_type
               uiKlines = FALSE
               quiet = FALSE)
}
\arguments{
\item{pair}{Character, trading pair, e.g. "BTCUSDT".}

\item{api}{Character, reference API. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://binance-docs.github.io/apidocs/spot/en/#kline-candlestick-data}{Spot API}.
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#kline-candlestick-data}{Futures USD-m API}.
\item \code{"dapi"}: for \href{https://binance-docs.github.io/apidocs/delivery/en/#kline-candlestick-data}{Futures COIN-m API}.
\item \code{"eapi"}: for \href{https://binance-docs.github.io/apidocs/voptions/en/#kline-candlestick-data}{Options API}.
}}

\item{interval}{Character, default is \code{"1d"}. Time interval for Klines data. Available intervals are:
\itemize{
\item \code{secondly}: \code{"1s"}, available only for \code{api = "spot"}.
\item \code{minutely}: \code{"1m"}, \code{"3m"}, \code{"5m"}, \code{"15m"} and \code{"30m"};
\item \code{hourly}: \code{"1h"}, \code{"2h"}, \code{"4h"}, \code{"6h"}, \code{"8h"} and \code{"12h"};
\item \code{daily}: \code{"1d"} and \code{"3d"};
\item \code{weekly}: \code{"1w"};
\item \code{monthly}: \code{"1M"}.
}}

\item{from}{Character or \code{"\link[=POSIXt-class]{POSIXt}"} object. Start time for historical data.
Default is \code{missing} and will be used as start date \code{Sys.time()-lubridate::days(1)}.}

\item{to}{Character or \code{"\link[=POSIXt-class]{POSIXt}"} object. End time for historical data.
Default is \code{missing} and will be used as end date \code{"\link[=Sys.time]{Sys.time()}"}.}

\item{contract_type}{Character, used only if \code{api = "dapi"}. Available contract type are:
\itemize{
\item \code{"perpetual"}: perpetual futures.
\item \code{"current_quarter"}: futures contracts with maturity in the current quarter.
\item \code{"next_quarter"}: futures contracts with maturity in the next quarter.
}}

\item{uiKlines}{Logical, if \code{TRUE} return data in UIklines format. Default is \code{FALSE}.}

\item{quiet}{Logical, if \code{TRUE} suppress informational and warnings. Default is \code{FALSE}}
}
\value{
A tibble (data frame) with the following 13 columns:
\itemize{
\item \code{date}: \code{"\link[=POSIXt-class]{POSIXt}"}, the opening date of the candle.
\item \code{date_close}: \code{"\link[=POSIXt-class]{POSIXt}"}, the closing date of the candle.
\item \code{market}: Character, selected API.
\item \code{pair}: Character, selected trading pair.
\item \code{open}: Numeric, open price or price in \code{date}.
\item \code{high}: Numeric, highest price from \code{date} up to \code{date_close}.
\item \code{low}: Numeric, lowest price from \code{date} up to \code{date_close}.
\item \code{close}: Numeric, close price or price in \code{date_close}.
\item \code{volume}: Numeric, volume in asset value.
\item \code{volume_quote}: Numeric, volume in quote value.
\item \code{trades}: Numeric, number of trades from \code{date} up to \code{date_close}.
\item \code{taker_buy}: Numeric, buy volume in asset value.
\item \code{taker_buy_quote}: Numeric, buy volume in quote value.
}
}
\description{
Get candlestick data for a trading pair with Binance API.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Get 1-hour OHLC data for BTCUSDT in the spot market
binance_klines(api = "spot", pair = "BTCUSDT", interval = "1h")

# Get 30-minute OHLC data for BNBUSDT in the perpetual market
binance_klines(api = "fapi", pair = "BNBUSDT", interval = "30m")

# Get 15-minute OHLC data for BTCUSD_PERP in the futures market
binance_klines(api = "dapi", pair = "BTCUSD_PERP", interval = "15m")

# Get 1-hour OHLC data for a put option on BTCUSDT with a strike of 30000 and maturity on 2024-06-28
binance_klines(api = "eapi", pair = "BTC-240628-30000-P", interval = "1h")

}
