% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_klines.R
\name{binance_klines}
\alias{binance_klines}
\title{Get Klines Data from Binance API}
\usage{
binance_klines(
  pair,
  interval,
  from,
  to,
  api = "spot",
  contract_type,
  uiKlines = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{pair}{Character, trading pair, e.g. "BTCUSDT".}

\item{interval}{Character, time interval for Klines data. Available intervals are:
\itemize{
\item \code{secondly}: "1s";
\item \code{minutely}: "1m", "3m", "5m", "15m" and "30m";
\item \code{hourly}: "1h", "2h", "4h", "6h", "8h" and "12h";
\item \code{daily}: "1d" and "3d";
\item \code{weekly}: "1w";
\item \code{monthly}: "1M".
}}

\item{from}{Character or an object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the start time for historical data.
Default is \code{NULL} and will be used as start date \code{Sys.time()-lubridate::days(1)}.}

\item{to}{Character or an object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the end time for historical data.
Default is \code{NULL} and will be used as end date \code{Sys.time()}.}

\item{api}{Character, reference API. Available options are:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#kline-candlestick-data}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#kline-candlestick-data}{Futures USD-m API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#kline-candlestick-data}{Futures COIN-m API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#kline-candlestick-data}{Options API}.
}}

\item{contract_type}{Character, used only if \code{api = "dapi"}. Available contract type are:
\itemize{
\item "perpetual": for perpetual futures.
\item "current_quarter": for futures with a maturity in the current quarter.
\item "next_quarter": for futures with a maturity in the next quarter.
}}

\item{uiKlines}{Logical, if \code{TRUE} return data in UIklines format. Default is \code{FALSE}.}

\item{quiet}{Logical, if \code{TRUE} suppress informational and warnings. Default is \code{FALSE}.}
}
\value{
A tibble (data frame) with the following 13 columns:
\itemize{
\item \code{date}: POSIXct, the opening date of the candle.
\item \code{date_close}: POSIXct, the closing date of the candle.
\item \code{market}: Character, the selected API.
\item \code{pair}: Character, the selected trading pair.
\item \code{open}: Numeric, the opening price.
\item \code{high}: Numeric, the highest price.
\item \code{low}: Numeric, the lowest price.
\item \code{close}: Numeric, the closing price.
\item \code{volume}: Numeric, the volume of the asset in asset value.
\item \code{volume_quote}: Numeric, the volume of the asset in quote value.
\item \code{trades}: Numeric, the number of trades.
\item \code{taker_buy}: Numeric, the buy volume of the asset in asset value.
\item \code{taker_buy_quote}: Numeric, the buy volume of the asset in quote value.
}
}
\description{
Retrieve Klines (candlestick) data for a specific trading pair from the Binance API. This function supports various Binance API types, including spot, futures, and options.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{
# Get 1-hour OHLC data for BTCUSDT in the spot market
binance_klines(api = "spot", pair = "BTCUSDT", interval = "1h")

# Get 30-minute OHLC data for BNBUSDT in the perpetual market
binance_klines(api = "fapi", pair = "BNBUSDT", interval = "30m")

# Get 15-minute OHLC data for BTCUSD_PERP in the futures market
binance_klines(api = "dapi", pair = "BTCUSD_PERP", interval = "15m")

# Get 1-hour OHLC data for a put option on BTCUSDT with a strike of 30000 and maturity on 2024-06-28
binance_klines(api = "eapi", pair = "BTC-240628-30000-P", interval = "1h")

}
