% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_klines.R
\name{binance_klines}
\alias{binance_klines}
\title{Get Klines Data from Binance API}
\usage{
binance_klines(
  pair,
  interval,
  from,
  to,
  api = "spot",
  contract_type,
  uiKlines = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{pair}{Character, the trading pair of interest, e.g., "BTCUSDT".}

\item{interval}{Character, the time interval for Klines data.
Acceptable intervals include "1s", "1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d", "3d", "1w", and "1M".}

\item{from}{Character or an object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the start time for data retrieval. If left unspecified (NULL), the default is Sys.Date()-1.}

\item{to}{Character or an object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the end time for data retrieval. If left unspecified (NULL), the default is Sys.Date().}

\item{api}{Character, specifying the reference API. Available options include:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#kline-candlestick-data}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#kline-candlestick-data}{Futures USD-M API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#kline-candlestick-data}{Futures Coin-M API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#kline-candlestick-data}{Options API}.
}}

\item{contract_type}{Character, only used if \code{api = "dapi"}. Choose from:
\itemize{
\item "perpetual": For perpetual futures.
\item "current_quarter": For futures with a maturity in the current quarter.
\item "next_quarter": For futures with a maturity in the next quarter.
}}

\item{uiKlines}{Logical, indicating whether to return data in UI Klines format.}

\item{quiet}{Logical, suppress informational messages if TRUE.}
}
\value{
A tibble (data frame) with the following 13 columns:
\itemize{
\item \code{date}: POSIXct, the opening date of the candle.
\item \code{date_close}: POSIXct, the closing date of the candle.
\item \code{market}: Character, the selected API.
\item \code{pair}: Character, the selected trading pair.
\item \code{open}: Numeric, the opening price.
\item \code{high}: Numeric, the highest price.
\item \code{low}: Numeric, the lowest price.
\item \code{close}: Numeric, the closing price.
\item \code{volume}: Numeric, the volume of the asset in asset value.
\item \code{volume_quote}: Numeric, the volume of the asset in quote value.
\item \code{trades}: Numeric, the number of trades.
\item \code{taker_buy}: Numeric, the buy volume of the asset in asset value.
\item \code{taker_buy_quote}: Numeric, the buy volume of the asset in quote value.
}
}
\description{
Retrieve Klines (candlestick) data for a specific trading pair from the Binance API. This function supports various Binance API types, including spot, futures, and options.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{
# Get 1-hour OHLC data for BTCUSDT in the spot market
binance_klines(api = "spot", pair = "BTCUSDT", interval = "1h")

# Get 30-minute OHLC data for BNBUSDT in the perpetual market
binance_klines(api = "fapi", pair = "BNBUSDT", interval = "30m")

# Get 15-minute OHLC data for BTCUSD_PERP in the futures market
binance_klines(api = "dapi", pair = "BTCUSD_PERP", interval = "15m")

# Get 1-hour OHLC data for a put option on BTCUSDT with a strike of 30000 and maturity on 2024-06-28
binance_klines(api = "eapi", pair = "BTC-240628-30000-P", interval = "1h")

}
