% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_last_trades.R
\name{binance_last_trades}
\alias{binance_last_trades}
\title{Retrieve Recent Trades Data}
\usage{
binance_last_trades(pair, 
                    api, 
                    quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://binance-docs.github.io/apidocs/spot/en/#recent-trades-list}{spot API}.
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#recent-trades-list}{futures USD-m API}.
\item \code{"dapi"}: for \href{https://binance-docs.github.io/apidocs/delivery/en/#recent-trades-list}{futures COIN-m API}.
\item \code{"eapi"}: for \href{https://binance-docs.github.io/apidocs/voptions/en/#recent-trades-list}{options API}.
}}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[=data.frame-class]{data.frame}} with 7 columns:
\itemize{
\item \code{date}: \code{"\link[=POSIXt-class]{POSIXt}"}, trade execution date.
\item \code{market}: Character, selected API.
\item \code{pair}: Character, trading pair.
\item \code{price}: Numeric, trade price.
\item \code{quantity}: Numeric, trade quantity.
\item \code{side}: Character, trade side. Can be \code{"BUY"} or \code{"SELL"}.
\item \code{trade_id}: Integer, trade id.
}
}
\description{
Get the last 1000 trades for a trading pair.
}
\examples{

# Get last 1000 trades for BTCUSDT
binance_last_trades(pair = "BTCUSDT", api = "spot")
binance_last_trades(pair = "BTCUSDT", api = "fapi")

# Get last 1000 trades for BTCUSD_PERP
binance_last_trades(pair = "BTCUSD_PERP", api = "dapi")

# Get last 1000 trades for a put option on BTC
binance_last_trades(pair = "BTC-240628-30000-P", api = "eapi")

}
