% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_last_trades.R
\name{binance_last_trades}
\alias{binance_last_trades}
\title{Retrieve Recent Trades Data}
\usage{
binance_last_trades(pair, api = "spot", quiet = FALSE)
}
\arguments{
\item{pair}{Character, specifying the trading pair of interest, e.g., "BTCUSDT".}

\item{api}{Character, specifying the reference API. Available options include:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#recent-trades-list}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#recent-trades-list}{Futures USD-M API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#recent-trades-list}{Futures Coin-M API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#recent-trades-list}{Options API}.
}}

\item{quiet}{Logical, specifying whether to suppress informational messages. Default is FALSE.}
}
\value{
A tibble (data frame) with the following 7 columns:
\itemize{
\item \code{trade_id}: Column containing the trade ID.
\item \code{date}: Datetime, representing the trade execution date.
\item \code{pair}: Column specifying the selected trading pair.
\item \code{market}: Column indicating the selected API.
\item \code{price}: Column showing the trade price.
\item \code{quantity}: Column representing the trade quantity.
\item \code{side}: Character, indicating whether the trade was a "SELL" or "BUY."
}
}
\description{
Obtain the most recent 1000 trades for a specified trading pair. This function allows you to access the latest trades that have occurred in a specific market.
}
\examples{

# Example: Get the most recent trades for the BTCUSDT trading pair in the Spot market.
binance_last_trades(pair = "BTCUSDT", api = "spot")

# Example: Get the most recent trades for the BTCUSDT trading pair in the Futures USD-M market.
binance_last_trades(pair = "BTCUSDT", api = "fapi")

# Example: Get the most recent trades for the BTCUSD_PERP trading pair in the Futures COIN-M market.
binance_last_trades(pair = "BTCUSD_PERP", api = "dapi")

}
