% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_last_trades.R
\name{binance_last_trades}
\alias{binance_last_trades}
\title{Retrieve Recent Trades Data}
\usage{
binance_last_trades(pair, api, quiet = FALSE)
}
\arguments{
\item{pair}{Character, trading pair, e.g. "BTCUSDT".}

\item{api}{Character, reference API. Available options are:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#recent-trades-list}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#recent-trades-list}{Futures USD-m API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#recent-trades-list}{Futures COIN-m API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#recent-trades-list}{Options API}.
}}

\item{quiet}{Logical, if \code{TRUE} suppress informational and warnings. Default is \code{FALSE}.}
}
\value{
A tibble with 7 columns:
\itemize{
\item \code{date}: Datetime, trade execution date.
\item \code{market}: Character, selected API.
\item \code{pair}: Character, selected pair.
\item \code{price}: Numeric, trade price.
\item \code{quantity}: Numeric, trade quantity.
\item \code{side}: Character, trade side. Can be "SELL" or "BUY".
\item \code{trade_id}: Integer, trade Id.
}
}
\description{
Get the most recent 1000 trades for a specified trading pair.
}
\examples{

# Get the last 1000 trades for BTCUSDT in spot market
binance_last_trades(pair = "BTCUSDT", api = "spot")

# Get the last 1000 trades for BTCUSDT in USD-M market
binance_last_trades(pair = "BTCUSDT", api = "fapi")

# Get the last 1000 trades for BTCUSDT in COIN-M market
binance_last_trades(pair = "BTCUSD_PERP", api = "dapi")

# Get the last 1000 trades for BTCUSDT in COIN-M market
binance_last_trades(pair = "BTC-240628-30000-P", api = "eapi")

}
