% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws_live.R
\name{binance_live_klines}
\alias{binance_live_klines}
\title{binance_live_klines}
\usage{
binance_live_klines(pair = "BTCUSDT", interval = "1m", api, data = NULL)
}
\arguments{
\item{pair}{Character, the trading pair of interest, e.g., "BTCUSDT".}

\item{interval}{Character, the time interval for Klines data.
Acceptable intervals include "1s", "1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d", "3d", "1w", and "1M".}

\item{api}{Character, specifying the reference API. Available options include:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#kline-candlestick-data}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#kline-candlestick-data}{Futures USD-M API}.
}}

\item{data}{initial data}
}
\description{
live candlestick
}
