% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws_live.R
\name{binance_live_trades}
\alias{binance_live_trades}
\title{binance_live_trades}
\usage{
binance_live_trades(pair = "BTCUSDT", api)
}
\arguments{
\item{pair}{Character, the trading pair of interest, e.g., "BTCUSDT".}

\item{api}{Character, specifying the reference API. Available options include:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#kline-candlestick-data}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#kline-candlestick-data}{Futures USD-M API}.
}}
}
\description{
live trades
}
