% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{binance_mytrades}
\alias{binance_mytrades}
\title{Get all trades on the Binance account}
\usage{
binance_mytrades(symbol)
}
\arguments{
\item{symbol}{string}
}
\value{
data.table
}
\description{
Get all trades on the Binance account
}
\examples{
\dontrun{
binance_mytrades('ETH')
}
}
