% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_mytrades}
\alias{binance_mytrades}
\title{Get trades for a specific symbol on the Binance account}
\usage{
binance_mytrades(symbol, limit, from_id, start_time, end_time)
}
\arguments{
\item{symbol}{string}

\item{limit}{optional int number of trades to fetch}

\item{from_id}{optional trade id to fetch from}

\item{start_time}{optional POSIX timestamp}

\item{end_time}{optional POSIX timestamp}
}
\value{
data.table
}
\description{
Get trades for a specific symbol on the Binance account
}
\examples{
\dontrun{
binance_mytrades('ARKETH')
binance_mytrades(c('ARKBTC', 'ARKETH'))
}
}
