% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_new_order}
\alias{binance_new_order}
\title{Open new order on the Binance account}
\usage{
binance_new_order(symbol, side = c("BUY", "SELL"), type = c("LIMIT",
  "MARKET", "STOP_LOSS", "STOP_LOSS_LIMIT", "TAKE_PROFIT",
  "TAKE_PROFIT_LIMIT", "LIMIT_MAKER"), timeInForce = c("GTC", "IOC",
  "FOK"), quantity, price, icebergQty, test = TRUE)
}
\arguments{
\item{symbol}{string}

\item{side}{enum}

\item{type}{enum}

\item{timeInForce}{enum}

\item{quantity}{number}

\item{price}{number}

\item{icebergQty}{number}

\item{test}{bool}
}
\value{
data.table
}
\description{
Open new order on the Binance account
}
\examples{
\dontrun{
binance_mytrades('ARKETH')
binance_mytrades(c('ARKBTC', 'ARKETH'))
}
}
