% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_new_order}
\alias{binance_new_order}
\title{Open new order on the Binance account}
\usage{
binance_new_order(
  symbol,
  side = c("BUY", "SELL"),
  type = c("LIMIT", "MARKET", "STOP_LOSS", "STOP_LOSS_LIMIT", "TAKE_PROFIT",
    "TAKE_PROFIT_LIMIT", "LIMIT_MAKER"),
  time_in_force = c("GTC", "IOC", "FOK"),
  quantity,
  quote_order_qty,
  price,
  stop_price,
  iceberg_qty,
  test = TRUE
)
}
\arguments{
\item{symbol}{string}

\item{side}{enum}

\item{type}{enum}

\item{time_in_force}{optional enum}

\item{quantity}{number The quantity of the base asset to trade. See details for an example.}

\item{quote_order_qty}{number The quantity of the quote asset to trade. See details for an example.}

\item{price}{optional number}

\item{stop_price}{optional number}

\item{iceberg_qty}{optional number}

\item{test}{bool}
}
\value{
data.table
}
\description{
Open new order on the Binance account
}
\details{
If trading `BTCUSDT` for example using `quantity` allows specification of the quantity of `BTC` of the trade, while `quote_order_qty` allows specification of the USDT value of the trade.
}
\examples{
\dontrun{
binance_new_order('ARKETH', side = 'BUY', type = 'MARKET', quantity = 1)
binance_new_order('ARKBTC', side = 'BUY', type = 'LIMIT', quantity = 1,
                  price = 0.5, time_in_force = 'GTC')
}

}
