% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_open_interest.R
\name{binance_open_interest}
\alias{binance_open_interest}
\title{Retrieve Present Open Interest Data}
\usage{
binance_open_interest(
  pair,
  api = "fapi",
  expiration = Sys.Date(),
  quiet = FALSE
)
}
\arguments{
\item{pair}{Character, specifying the trading pair of interest, e.g., "BTCUSDT".}

\item{api}{Character, specifying the reference API. Available options include:
\itemize{
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#open-interest}{Futures USD-M API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#open-interest}{Futures Coin-M API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#open-interest}{Options API}.
}}

\item{expiration}{POSIXct, specifying the expiration date for options contracts. This parameter is used when \code{api = "eapi"}.
Default is NULL, which represents the most recent data.
You can provide a POSIXct object, e.g., Sys.Date() + 1 or a specific date and time in the future.}
}
\value{
A tibble (data frame) object containing the current open interest data, including date, market, pair, and open interest values.
}
\description{
Obtain the current open interest data for a specific trading pair. This function allows you to access the real-time open interest information.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Example: Retrieve the current open interest data for the "BTCUSDT" trading pair in the Futures USD-M market.
binance_open_interest(pair = "BTCUSDT", api = "fapi")

# Example: Retrieve the current open interest data for the "BTCUSD_PERP" trading pair in the Futures Coin-M market.
binance_open_interest(pair = "BTCUSD_PERP", api = "dapi")

# Example: Retrieve the current open interest data for options contracts on "BTC" with a specific expiration date.
binance_open_interest(pair = "BTC", api = "eapi", expiration = Sys.Date() + 1)

}
