% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_open_interest.R
\name{binance_open_interest}
\alias{binance_open_interest}
\title{Current Open Interest Data}
\usage{
binance_open_interest(pair, api, expiration = Sys.Date(), quiet = FALSE)
}
\arguments{
\item{pair}{Character, trading pair, e.g. "BTCUSDT".}

\item{api}{Character, reference API. Available options are:
\itemize{
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#open-interest}{Futures USD-m API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#open-interest}{Futures Coin-m API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#open-interest}{Options API}.
}}

\item{expiration}{POSIXct, specifying the expiration date for options contracts. This parameter is used when \code{api = "eapi"}.
Default is NULL, which represents the most recent data.
You can provide a POSIXct object, e.g., Sys.Date() + 1 or a specific date and time in the future.}

\item{quiet}{Logical, if \code{TRUE} suppress informational and warnings. Default is \code{FALSE}.}
}
\value{
A tibble with 4 columns:
\itemize{
\item \code{date}: Datetime, observation date.
\item \code{market}: Character, selected API.
\item \code{pair}: Character, selected pair.
\item \code{open_interest}: Numeric, open interest in base currency.
}
}
\description{
Get the current open interest data for a specific trading pair.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Get the open interest data for "BTCUSDT" in USD-M market
binance_open_interest(pair = "BTCUSDT", api = "fapi")

# Get the open interest data for "BTCUSD_PERP" in COIN-M market
binance_open_interest(pair = "BTCUSD_PERP", api = "dapi")

# Get the open interest data for options on BTC
binance_open_interest(pair = "BTC", api = "eapi", expiration = Sys.Date() + 1)

}
