% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_open_interest.R
\name{binance_open_interest}
\alias{binance_open_interest}
\title{Current Open Interest Data}
\usage{
binance_open_interest(pair, api, expiration, quiet = FALSE)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"fapi"}. Available options are:
\itemize{
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#open-interest}{futures USD-m API}.
\item \code{"dapi"}: for \href{https://binance-docs.github.io/apidocs/delivery/en/#open-interest}{futures Coin-m API}.
\item \code{"eapi"}: for \href{https://binance-docs.github.io/apidocs/voptions/en/#open-interest}{options API}.
}}

\item{expiration}{Character or \code{\link[=POSIXt-class]{POSIXt}} object. Used only if \code{api = "eapi"}.
Expiration date for options contracts. If it is \code{missing}, the default, will be used \code{\link[=Sys.time]{Sys.time()}}.}

\item{quiet}{Logical. Default is \code{FALSE}. If \code{TRUE} suppress messages and warnings.}
}
\value{
A \code{\link[=data.frame-class]{data.frame}} with 4 columns:
\itemize{
\item \code{date}: \code{\link[=POSIXt-class]{POSIXt}}, observation date.
\item \code{market}: Character, API.
\item \code{pair}: Character, trading pair.
\item \code{open_interest}: Numeric, open interest in base currency.
}
}
\description{
Get the current open interest data for a specific trading pair.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Get the open interest data for BTCUSDT
binance_open_interest(pair = "BTCUSDT", api = "fapi")

# Get the open interest data for BTCUSD_PERP
binance_open_interest(pair = "BTCUSD_PERP", api = "dapi")

# Get the open interest data for options on BTC
binance_open_interest(pair = "BTC", api = "eapi", expiration = Sys.Date() + 1)

}
