% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_open_interest_hist.R
\name{binance_open_interest_hist}
\alias{binance_open_interest_hist}
\title{Retrieve Historical Open Interest Data}
\usage{
binance_open_interest_hist(
  pair,
  api = "fapi",
  interval = "1d",
  from = NULL,
  to = NULL,
  contract_type = "all",
  quiet = FALSE
)
}
\arguments{
\item{pair}{Character, specifying the trading pair of interest, e.g., "BTCUSDT".}

\item{api}{Character, specifying the reference API. Default is "fapi". Available options include:
\itemize{
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#open-interest}{Futures USD-M API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#open-interest}{Futures Coin-M API}.
}}

\item{interval}{Character, specifying the time interval for open interest data. Default is "1d". Available intervals include "1d", "3d", "1w", and more.}

\item{from}{Character or POSIXct, specifying the start time for data retrieval. Default is NULL, which represents the earliest available data. You can provide a character or a POSIXct object, e.g., "2023-01-01" or a specific date and time.}

\item{to}{Character or POSIXct, specifying the end time for data retrieval. Default is NULL, which represents the most recent available data. You can provide a character or a POSIXct object, e.g., "2023-12-31" or a specific date and time.}
}
\value{
A tibble (data frame) object containing open interest data, including date, market, pair, and open interest values over time.
}
\description{
Obtain historical open interest data for a specific trading pair. This function allows you to access the open interest information over time.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Retrieve historical open interest for "BTCUSD" in USD-M market
binance_open_interest_hist(pair = "BTCUSDT", api = "fapi", interval = "1d", from = "2023-01-01")

# Retrieve historical open interest for "BTCUSD" in Coin-M market
binance_open_interest_hist(pair = "BTCUSD", api = "dapi", interval = "1d", from = "2023-01-01")

}
