% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ping.R
\name{binance_ping}
\alias{binance_ping}
\title{Test Connectivity to Binance REST API}
\usage{
binance_ping(api)
}
\arguments{
\item{api}{Character, reference API. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://binance-docs.github.io/apidocs/spot/en/#test-connectivity}{Spot API}.
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#test-connectivity}{Futures USD-m API}.
\item \code{"dapi"}: for \href{https://binance-docs.github.io/apidocs/delivery/en/#test-connectivity}{Futures COIN-m API}.
\item \code{"eapi"}: for \href{https://binance-docs.github.io/apidocs/voptions/en/#test-connectivity}{Options API}.
}}
}
\value{
\code{TRUE} if the connection was successful, otherwise \code{FALSE} if the connection failed.
}
\description{
Check the connectivity to the Binance API to ensure it is working.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Test connection to Spot API
binance_ping("spot")

# Test connection to Futures USD-m API
binance_ping("fapi")

# Test connection to Futures Coin-m API
binance_ping("dapi")

# Test connection to Options API
binance_ping("eapi")

}
