% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_query}
\alias{binance_query}
\title{Request the Binance API}
\usage{
binance_query(endpoint, method = c("GET", "POST"), params = list(),
  body = FALSE, sign = FALSE, retry = method == "GET")
}
\arguments{
\item{endpoint}{string}

\item{method}{HTTP request method}

\item{params}{URL parameters provided as a list}

\item{body}{body of the request}

\item{sign}{if signature required}

\item{retry}{allow retrying the query on failure}
}
\value{
R object
}
\description{
Request the Binance API
}
\keyword{internal}
