% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binance_query}
\alias{binance_query}
\title{Request the Binance API}
\usage{
binance_query(endpoint, method = "GET", params = list(), sign = FALSE,
  retry = method == "GET")
}
\arguments{
\item{endpoint}{string}

\item{method}{HTTP request method}

\item{params}{list}

\item{sign}{if signature required}

\item{retry}{allow retrying the query on failure}
}
\value{
R object
}
\description{
Request the Binance API
}
\keyword{internal}
