% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_ticker_all_prices}
\alias{binance_ticker_all_prices}
\title{Get latest Binance conversion rates and USD prices on all symbol pairs}
\usage{
binance_ticker_all_prices()
}
\value{
data.table
}
\description{
Get latest Binance conversion rates and USD prices on all symbol pairs
}
