% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_ticker_price}
\alias{binance_ticker_price}
\title{Get last price for a symbol or all symbols}
\usage{
binance_ticker_price(symbol)
}
\arguments{
\item{symbol}{optional string}
}
\value{
data.table
}
\description{
Get last price for a symbol or all symbols
}
