% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_ticks}
\alias{binance_ticks}
\title{Get tick data from Binance}
\usage{
binance_ticks(symbol, from_id, start_time, end_time, limit)
}
\arguments{
\item{symbol}{string}

\item{from_id}{optional number}

\item{start_time}{optional POSIX timestamp}

\item{end_time}{optional POSIX timestamp}

\item{limit}{optional int}
}
\value{
data.table
}
\description{
Get tick data from Binance
}
\examples{
\dontrun{
binance_ticks('ETHUSDT')
binance_ticks('ETHUSDT', start_time = '2018-01-01 00:00:00', end_time = '2018-01-01 01:00:00')
}
}
