% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_ticks}
\alias{binance_ticks}
\title{Get tick data from Binance}
\usage{
binance_ticks(symbol, start_time, end_time, limit = 500)
}
\arguments{
\item{symbol}{string}

\item{start_time}{POSIX timestamp}

\item{end_time}{POSIX timestamp}

\item{limit}{int}
}
\value{
data.table
}
\description{
Get tick data from Binance
}
\examples{
\dontrun{
binance_ticks('ETHUSDT')
binance_ticks('ETHUSDT', start_time = as.POSIXct('2018-01-01'), end_time = as.POSIXct('2018-01-08'))
}
}
