% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_time.R
\name{binance_time}
\alias{binance_time}
\title{Binance Server Time}
\usage{
binance_time(api, quiet = FALSE)
}
\arguments{
\item{api}{Character, reference API. Available options are:
\itemize{
\item \code{"spot"}: For \href{https://binance-docs.github.io/apidocs/spot/en/#check-server-time}{spot API};
\item \code{"fapi"}: For \href{https://binance-docs.github.io/apidocs/futures/en/#check-server-time}{futures USD-m API};
\item \code{"dapi"}: For \href{https://binance-docs.github.io/apidocs/delivery/en/#check-server-time}{futures COIN-m API};
\item \code{"eapi"}: For \href{https://binance-docs.github.io/apidocs/voptions/en/#check-server-time}{options API}.
}}

\item{quiet}{Logical, if \code{TRUE} suppress informational and warnings. Default is \code{FALSE}}
}
\value{
An object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the server's time for the reference API.
}
\description{
Obtain the current server time.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Get the server time for the Spot API.
binance_time("spot")

# Get the server time for the Futures USD-M API.
binance_time("fapi")

# Get the server time for the Futures Coin-M API.
binance_time("dapi")

# Get the server time for the Options API.
binance_time("eapi")

}
