% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_time.R
\name{binance_time}
\alias{binance_time}
\title{Binance Server Time}
\usage{
binance_time(api)
}
\arguments{
\item{api}{Character, reference API. Available options are:
\itemize{
\item "spot": For \href{https://binance-docs.github.io/apidocs/spot/en/#check-server-time}{Spot API}.
\item "fapi": For \href{https://binance-docs.github.io/apidocs/futures/en/#check-server-time}{Futures USD-m API}.
\item "dapi": For \href{https://binance-docs.github.io/apidocs/delivery/en/#check-server-time}{Futures COIN-m API}.
\item "eapi": For \href{https://binance-docs.github.io/apidocs/voptions/en/#check-server-time}{Options API}.
}}
}
\value{
An object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the server's time for the reference API.
}
\description{
Obtain the current server time.
}
\details{
The IP weight for this API call is 1, and the data source is memory.
}
\examples{

# Get the server time for the Spot API.
binance_time("spot")

# Get the server time for the Futures USD-M API.
binance_time("fapi")

# Get the server time for the Futures Coin-M API.
binance_time("dapi")

# Get the server time for the Options API.
binance_time("eapi")

}
