% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_trades}
\alias{binance_trades}
\title{Get last trades from Binance}
\usage{
binance_trades(symbol, limit)
}
\arguments{
\item{symbol}{string}

\item{limit}{optional int}
}
\value{
data.table
}
\description{
Get last trades from Binance
}
\examples{
\dontrun{
binance_ticks('ETHUSDT')
binance_ticks('ETHUSDT', start_time = as.POSIXct('2018-01-01'), end_time = as.POSIXct('2018-01-08'))
}
}
