% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance.R
\name{binance_trades}
\alias{binance_trades}
\title{Get last trades from Binance}
\usage{
binance_trades(symbol, limit)
}
\arguments{
\item{symbol}{string}

\item{limit}{optional int}
}
\value{
\code{data.table}
}
\description{
Get last trades from Binance
}
\examples{
\dontrun{
binance_trades('ETHUSDT')
binance_trades('ETHUSDT', limit = 1000)
}
}
