% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws.R
\name{binance_ws_fapi_socket}
\alias{binance_ws_fapi_socket}
\title{binance_ws_fapi_socket}
\usage{
binance_ws_fapi_socket(
  pair = "BTCUSDT",
  subscription = "kline",
  interval = "1m",
  contractType = NULL,
  update_speed = 1000,
  quiet = FALSE
)
}
\description{
fapi websocket
}
