% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_ws.R
\name{binance_ws_socket}
\alias{binance_ws_socket}
\title{Binance Websocket in R}
\usage{
binance_ws_socket(
  pair = "BTCUSDT",
  api = "spot",
  subscription = "kline",
  interval = "1m",
  contract_type = NULL,
  update_speed = 1000,
  nobs_max = NULL,
  quiet = FALSE
)
}
\arguments{
\item{pair}{Character. Trading pair, e.g. \code{"BTCUSDT"}.}

\item{api}{Character. Reference API. If it is \code{missing}, the default, will be used \code{"spot"}. Available options are:
\itemize{
\item \code{"spot"}: for \href{https://binance-docs.github.io/apidocs/spot/en/#kline-candlestick-data}{spot API}.
\item \code{"fapi"}: for \href{https://binance-docs.github.io/apidocs/futures/en/#kline-candlestick-data}{futures USD-m API}.
\item \code{"dapi"}: for \href{https://binance-docs.github.io/apidocs/delivery/en/#kline-candlestick-data}{futures COIN-m API}.
\item \code{"eapi"}: for \href{https://binance-docs.github.io/apidocs/voptions/en/#kline-candlestick-data}{options API}.
}}

\item{subscription}{Character}

\item{interval}{Character, the time interval for Klines data.
Acceptable intervals are "1s", "1m", "3m", "5m", "15m", "30m", "1h", "2h", "4h", "6h", "8h", "12h", "1d", "3d", "1w", and "1M".}

\item{contract_type}{Character, only used if \code{api = "dapi"}. Choose from:
\itemize{
\item \code{"perpetual"}: For perpetual futures.
\item \code{"current_quarter"}: For futures with a maturity in the current quarter.
\item \code{"next_quarter"}: For futures with a maturity in the next quarter.
}}

\item{update_speed}{time in milliseconds}

\item{nobs_max}{Integer, maximum number of observations to save.}

\item{quiet}{Logical, suppress informational messages if TRUE.}
}
\description{
Create a connection to Binance websocket API in R.
}
