% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candleChart.R
\name{candleChart}
\alias{candleChart}
\title{Candlestick Plot}
\usage{
candleChart(
  data,
  from = NULL,
  to = NULL,
  title = NULL,
  col_body_up = "green",
  col_body_dw = "red",
  col_wick = "black"
)
}
\arguments{
\item{data}{A data frame containing time-series data with columns: 'date', 'open', 'close', 'low', 'high', 'pair', and 'interval'.}

\item{from}{Character or an object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the start date for the plot.
If specified, only data from this date on wards will be included in the chart.}

\item{to}{Character or an object of class \code{"\link[=POSIXt-class]{POSIXt}"}, the end date for the plot.
If specified, only data up to this date will be included in the chart.}

\item{title}{Character, optional title for the chart.}

\item{col_body_up}{Character, color for candlestick bodies when the close price is higher than the open price. Default is "green".}

\item{col_body_dw}{Character, color for candlestick bodies when the close price is lower than the open price. Default is "red".}

\item{col_wick}{Character, color for candlestick wicks. Default is "black".}
}
\value{
A ggplot2 object representing the candlestick chart.
}
\description{
Create a candlestick chart to visualize price movements over a specified time frame.
}
