% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_utils.R
\name{coindata_firstloop}
\alias{coindata_firstloop}
\title{first round of coin dataset in backtest}
\usage{
coindata_firstloop(
  data,
  portval_loop1,
  min_momentum_count_loop1,
  min_trade_loop1,
  timeperiod_loop1,
  i_loop1
)
}
\arguments{
\item{fees}{}
}
\description{
first round of coin dataset in backtest
}
