% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klines.R
\name{get_klines}
\alias{get_klines}
\title{Get kline/candlestick data}
\usage{
get_klines(symbol, interval = c("1m", "3m", "5m", "15m", "30m", "1h", "2h",
  "4h", "6h", "8h", "12h", "1d", "3d", "1w", "1M"), limit = 500)
}
\arguments{
\item{symbol}{string}

\item{interval}{enum}

\item{limit}{int}
}
\value{
data.table
}
\description{
Get kline/candlestick data
}
