% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderBook.R
\name{OrderBook}
\alias{OrderBook}
\title{Order Book}
\usage{
OrderBook(
  data = NULL,
  min_price = NULL,
  max_price = NULL,
  levels = NULL,
  trades = NULL,
  as_datatable = FALSE
)
}
\arguments{
\item{data}{A tibble with at least 3 columns (price, quantity and side).}

\item{min_price}{Numeric, the minimum price for the aggregation of order book.}

\item{max_price}{Numeric, the maximum price for the aggregation of order book.}

\item{levels}{The number of levels for the aggregated order book.}

\item{trades}{A tibble with trades data.}

\item{as_datatable}{Logical, if \code{TRUE} return the order book as a datatable. Default is \code{FALSE}.}
}
\value{
A tibble with depth data in the specified price range.
}
\description{
Create and structure an order book from depth data.
}
