% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderBook.R
\name{OrderBook}
\alias{OrderBook}
\title{Create and Structure an Order Book}
\usage{
OrderBook(
  data = NULL,
  min_price = NULL,
  max_price = NULL,
  levels = 20,
  trades = NULL,
  as_datatable = FALSE
)
}
\arguments{
\item{data}{A tibble containing the order book dataset.}

\item{min_price}{The minimum price to consider for the order book.}

\item{max_price}{The maximum price to consider for the order book.}

\item{levels}{The number of levels to include in the order book.}

\item{trades}{A tibble containing the trades dataset to be integrated with the order book.}

\item{as_datatable}{Logical, specifying whether to return the structured order book as a data table (data frame). Default is FALSE.}
}
\value{
A structured tibble representing the order book with the specified price range and levels.
}
\description{
The \code{OrderBook} function is used to create and structure an existing order book dataset,
allowing you to specify the minimum and maximum price range and the number of levels to include.
}
