% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderBook.R
\name{OrderBook}
\alias{OrderBook}
\title{Order Book from Depth}
\usage{
OrderBook(
  data = NULL,
  min_price = NULL,
  max_price = NULL,
  levels = 20,
  trades = NULL,
  as_datatable = FALSE
)
}
\arguments{
\item{data}{A tibble containing the order book dataset.}

\item{min_price}{The minimum price to consider in the order book.}

\item{max_price}{The maximum price to consider in the order book.}

\item{levels}{The number of levels for the order book.}

\item{trades}{A tibble.}

\item{as_datatable}{Logical, specifying whether to return the structured order book as a data table (data frame). Default is FALSE.}
}
\value{
A structured tibble representing the order book with the specified price range and levels.
}
\description{
Create and structure an order book from depth data,
}
