% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderBook.R
\name{OrderBook}
\alias{OrderBook}
\title{OrderBook}
\usage{
OrderBook(
  data = NULL,
  min_price = NULL,
  max_price = NULL,
  levels = 20,
  trades = NULL,
  as_datatable = FALSE
)
}
\arguments{
\item{data}{tibble, order book dataset}

\item{min_price}{minimum price}

\item{max_price}{maximum price}

\item{levels}{number of levels}

\item{trades}{tibble, trades dataset}

\item{as_datatable}{logical}
}
\value{
a tibble
}
\description{
create and/or structure an existent order book
}
