% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query}
\alias{query}
\title{Request the Binance API}
\usage{
query(base, path, method, params = list(), body = NULL,
  config = config(), retry = method == "GET", retries = 0)
}
\arguments{
\item{base}{URL}

\item{path}{string}

\item{method}{HTTP request method}

\item{params}{URL parameters provided as a list}

\item{body}{body of the request}

\item{config}{httr::config}

\item{retry}{allow retrying the query on failure}

\item{retries}{internal counter of previous retries}
}
\value{
R object
}
\description{
Request the Binance API
}
\keyword{internal}
