% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdm.R
\name{usdm_v1_open_orders}
\alias{usdm_v1_open_orders}
\title{Get all open USDM orders on a symbol.}
\usage{
usdm_v1_open_orders(symbol)
}
\arguments{
\item{symbol}{string}
}
\value{
data.table
}
\description{
Get all open USDM orders on a symbol.
}
